% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae.R
\docType{data}
\name{adae}
\alias{adae}
\title{Example ADAE data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 21 columns.}
\usage{
data(adae)
}
\description{
Example ADAE data
}
\examples{
data(adae)
}
\keyword{datasets}
