% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaette.R
\docType{data}
\name{adaette}
\alias{adaette}
\title{Example ADAETTE data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1200 rows and 25 columns.}
\usage{
data(adaette)
}
\description{
Example ADAETTE data
}
\examples{
data(adaette)
}
\keyword{datasets}
