% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcm.R
\docType{data}
\name{adcm}
\alias{adcm}
\title{Example ADCM data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 21 columns.}
\usage{
data(adcm)
}
\description{
Example ADCM data
}
\examples{
data(adcm)
}
\keyword{datasets}
