% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlb.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{Example ADLB data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 33 columns.}
\usage{
data(adlb)
}
\description{
Example ADLB data
}
\examples{
data(adlb)
}
\keyword{datasets}
