% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrs.R
\docType{data}
\name{adrs}
\alias{adrs}
\title{Example ADRS data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 23 columns.}
\usage{
data(adrs)
}
\description{
Example ADRS data
}
\examples{
data(adrs)
}
\keyword{datasets}
