% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtte.R
\docType{data}
\name{adtte}
\alias{adtte}
\title{Example ADTTE data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1200 rows and 24 columns.}
\usage{
data(adtte)
}
\description{
Example ADTTE data
}
\examples{
data(adtte)
}
\keyword{datasets}
