% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advs.R
\docType{data}
\name{advs}
\alias{advs}
\title{Example ADVS data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 33 columns.}
\usage{
data(advs)
}
\description{
Example ADVS data
}
\examples{
data(advs)
}
\keyword{datasets}
