% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_metadata}
\alias{apply_metadata}
\title{Apply attributes to domains.}
\usage{
apply_metadata(df, filename, ..., ADSL = NULL)
}
\arguments{
\item{df}{data frame to which metadata are applied.}

\item{filename}{yaml file containing domain metadata.}

\item{ADSL}{logical to control merging of ADSL data to domain.}
}
\description{
Apply attributes to domains.
}
\examples{
 ADSL <- apply_metadata(ADSL, "ADSL.yml", seed = seed, ADSL = FALSE)
 ADAE <- apply_metadata(ADAE, "ADAE.yml", seed = NULL, ADSL = TRUE)

}
