% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadaette}
\alias{cadaette}
\title{Cached ADAETTE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3600 rows and 50 columns.
}
\usage{
data(cadaette)
}
\description{
Cached ADAETTE data generated with \code{seed = 1}
}
\keyword{datasets}
