% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadcm.R
\docType{data}
\name{cadcm}
\alias{cadcm}
\title{Cached ADCM data generated with \code{seed = 1}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 21 columns.}
\usage{
data(cadcm)
}
\description{
Cached ADCM data generated with \code{seed = 1}
}
\examples{
data(cadcm)
}
\keyword{datasets}
