% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{caddv}
\alias{caddv}
\title{Cached ADDV}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 103 rows and 55 columns.
}
\usage{
data(caddv)
}
\description{
Cached ADDV data generated with \code{seed = 1}
}
\keyword{datasets}
