% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadeg}
\alias{cadeg}
\title{Cached ADEG data generated with \code{seed = 1}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13600 rows and 66 columns.}
\usage{
data(cadeg)
}
\description{
Cached ADEG data generated with \code{seed = 1}
}
\keyword{datasets}
