% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadex}
\alias{cadex}
\title{Cached ADEX data generated with \code{seed = 1}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 45 columns.
}
\usage{
data(cadex)
}
\description{
Cached ADEX data generated with \code{seed = 1}
}
\keyword{datasets}
