% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadhy}
\alias{cadhy}
\title{Cached ADHY}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 52 columns.
}
\usage{
data(cadhy)
}
\description{
Cached ADHY data generated with \code{seed = 1}
}
\keyword{datasets}
