% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadlb.R
\docType{data}
\name{cadlb}
\alias{cadlb}
\title{Cached ADLB data generated with \code{seed = 1}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 33 columns.}
\usage{
data(cadlb)
}
\description{
Cached ADLB data generated with \code{seed = 1}
}
\examples{
data(cadlb)
}
\keyword{datasets}
