% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadmh}
\alias{cadmh}
\title{Cached ADMH data generated with \code{seed = 1}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 38 columns.}
\usage{
data(cadmh)
}
\description{
Cached ADMH data generated with \code{seed = 1}
}
\keyword{datasets}
