% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadpp}
\alias{cadpp}
\title{Cached ADPP}
\format{
An object of class \code{data.frame} with 26268 rows and 68 columns.
}
\usage{
data(cadpp)
}
\description{
Cached ADPP data generated with \code{seed = 1}
}
\keyword{datasets}
