% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadqlqc}
\alias{cadqlqc}
\title{Cached ADQLQC}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2869 rows and 50 columns.
}
\usage{
data(cadqlqc)
}
\description{
Cached ADQLQC data generated with \code{seed = 1}
}
\keyword{datasets}
