% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadqs}
\alias{cadqs}
\title{Cached ADQS}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14000 rows and 57 columns.
}
\usage{
data(cadqs)
}
\description{
Cached ADQS data generated with \code{seed = 1}
}
\keyword{datasets}
