% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadsl}
\alias{cadsl}
\title{Cached ADSL data generated with \code{seed = 1}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 31 columns.}
\usage{
data(cadsl)
}
\description{
Cached ADSL data generated with \code{seed = 1}
}
\keyword{datasets}
