% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadsub}
\alias{cadsub}
\title{Cached ADSUB}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 51 columns.
}
\usage{
data(cadsub)
}
\description{
Cached ADSUB data generated with \code{seed = 1}
}
\keyword{datasets}
