% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadtr}
\alias{cadtr}
\title{Cached ADTR data generated with \code{seed = 1}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 56 columns.
}
\usage{
data(cadtr)
}
\description{
Cached ADTR data generated with \code{seed = 1}
}
\keyword{datasets}
