% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadtte}
\alias{cadtte}
\title{Cached ADTTE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1600 rows and 52 columns.
}
\usage{
data(cadtte)
}
\description{
Cached ADTTE data generated with \code{seed = 1}
}
\keyword{datasets}
