% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadvs}
\alias{cadvs}
\title{Cached ADVS data generated with \code{seed = 1}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 58 columns.}
\usage{
data(cadvs)
}
\description{
Cached ADVS data generated with \code{seed = 1}
}
\keyword{datasets}
