% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{calc_scales}
\alias{calc_scales}
\title{Scale calculation for ADQLQC data}
\usage{
calc_scales(adqlqc1)
}
\arguments{
\item{adqlqc1}{prepared data generated from the prep_adqlqc function}
}
\value{
dataframe
}
\description{
Scale calculation for ADQLQC data
}
\examples{
\dontrun{
df <- calc_scales(adqlqc1)
}
}
