% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{comp_derv}
\alias{comp_derv}
\title{Completion/Compliance data calculation}
\usage{
comp_derv(dataset, percent, number)
}
\arguments{
\item{dataset}{adam dataset}

\item{percent}{Completion - Completed at least y percent of questions, 1 record per visit}

\item{number}{Completion - Completed at least x question(s), 1 record per visit}
}
\value{
data frame
}
\description{
Completion/Compliance data calculation
}
\examples{
\dontrun{
compliance_data <- comp_derv(ADQLQC, 80, 2)
}
}
