% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_region_all.R
\docType{data}
\name{country_region_all}
\alias{country_region_all}
\title{Country Region Codelist Dataset}
\format{
A data frame with variables:
\describe{
\item{REGION1}{Geographic region - long form}
\item{ACOUNTRY}{Country name - long form}
}
}
\source{
This medata was derived externally to this package.
}
\usage{
country_region_all
}
\description{
This dataset contains CDISC compliant metadata about countries and regions
frequently featured in clinical trial. These metadata are used to generate
countries and regions in ADSL.
}
\keyword{datasets}
