% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{derv_chgcat1}
\alias{derv_chgcat1}
\title{Calculate Change from Baseline Category 1}
\usage{
derv_chgcat1(dataset)
}
\arguments{
\item{dataset}{adam dataset}
}
\value{
dataframe
}
\description{
Calculate Change from Baseline Category 1
}
\examples{
\dontrun{
adqlqc <- derv_chgcat1(dataset = adqlqc)
}
}
\keyword{internal}
