% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{get_qs_data}
\alias{get_qs_data}
\title{Questionnaires EORTC QLQ-C30 V3.0 SDTM (QS)}
\usage{
get_qs_data(
  ADSL,
  visit_format = "CYCLE",
  n_assessments = 5L,
  n_days = 1L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(QSORRES = c(1234, 0.2), QSSTRESC = c(1234, 0.2))
)
}
\arguments{
\item{ADSL}{(\code{data.frame})\cr Subject-Level Analysis Dataset (ADSL).}

\item{visit_format}{(\code{character})\cr Type of visit. Options are "WEEK" and "CYCLE".}

\item{n_assessments}{(\code{integer})\cr Number of weeks or cycles.}

\item{n_days}{(\code{integer})\cr Number of days within cycle (only used if \code{visit_format} is "CYCLE").}

\item{lookup}{(\code{data.frame})\cr Additional parameters.}

\item{seed}{(\code{integer})\cr Seed for reproducible random number generation.}

\item{na_percentage}{(\code{proportion})\cr Default percentage of values to be replaced by \code{NA}.}

\item{na_vars}{(\code{list})\cr A named list where the name of each element is a column name of \code{ds}. Each
element of this list should be a numeric vector with two elements:
\itemize{
\item{\code{seed} (\code{integer})\cr The seed to be used for this element - can be \code{NA}.}
\item{\code{percentage} (\code{proportion})\cr Percentage of elements to be replaced with \code{NA}.
If \code{NA}, \code{na_percentage} is used as a default.}
}}
}
\value{
a dataframe with SDTM questionnaire data
}
\description{
Function for generating random Questionnaires SDTM domain
}
\examples{
\dontrun{
qs <- get_qs_data(adsl, n_assessments = 5L, seed = 1, na_percentage = 0.1)
qs
}
}
\keyword{internal}
