% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{get_questionnaire_sdtm}
\alias{get_questionnaire_sdtm}
\title{Generate SDTM QS from gdsr}
\usage{
get_questionnaire_sdtm(pub_nam = NULL, qscat)
}
\arguments{
\item{pub_nam}{gdsr publication name, default NULL}

\item{qscat}{Category of questionnaire}
}
\value{
dataframe
}
\description{
Generate SDTM QS from gdsr
}
\examples{
\dontrun{
eortc_qlq_c30 <- get_questionnaire_sdtm(qscat = "EORTC QLQ-C30 V3.0")
}
}
