% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{get_random_dates_between}
\alias{get_random_dates_between}
\title{Function for generating random dates between 2 dates}
\usage{
get_random_dates_between(from, to, visit_id)
}
\arguments{
\item{from}{start date/time variable}

\item{to}{end date/time variable}

\item{visit_id}{visit number variable}
}
\value{
data frame with new variable with random generated dates
}
\description{
Function for generating random dates between 2 dates
}
\examples{
\dontrun{
df <- df \%>\%
  group_by(USUBJID) \%>\%
  mutate(ADTM = get_random_dates_between(TRTSDTM, TRTEDTM, AVISITN))
df
}
}
\keyword{internal}
