% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{prep_adqlqc}
\alias{prep_adqlqc}
\title{Prepare ADaM ADQLQC data, adding PARAMCD to SDTM QS data}
\usage{
prep_adqlqc(df)
}
\arguments{
\item{df}{SDTM QS}
}
\value{
data frame
}
\description{
Prepare ADaM ADQLQC data, adding PARAMCD to SDTM QS data
}
\examples{
\dontrun{
adqlqc1 <- prep_adqlqc(df = qs)
adqlqc1
}
}
\keyword{internal}
