% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radcm.R
\name{radcm}
\alias{radcm}
\title{Concomitant Medication Analysis Dataset (ADCM)}
\usage{
radcm(ADSL, max_n_cms = 10, seed = NULL, lookup = NULL)
}
\arguments{
\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}

\item{max_n_cms}{The maximum number of concomitant medications per patient.}

\item{seed}{Seed for random number generation.}

\item{lookup}{Data.frame containing additional parameters.}
}
\value{
data.frame
}
\description{
Function for generating random Concomitant Medication Analysis Dataset for a given
Subject-Level Analysis Dataset
}
\examples{

ADSL <- radsl()
ADCM <- radcm(ADSL)

head(ADCM)

}
