% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raddv.R
\name{raddv}
\alias{raddv}
\title{Protocol Deviations Analysis Dataset (ADDV)}
\usage{
raddv(
  ADSL,
  max_n_dv = 3L,
  p_dv = 0.15,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(ASTDT = c(seed = 1234, percentage = 0.1), DVCAT = c(seed = 1234,
    percentage = 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{max_n_dv}{Maximum number of deviations per patient.}

\item{p_dv}{Probability of a patient to have protocol deviations.}

\item{lookup}{Data.frame containing additional parameters.}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached ADDV data \code{caddv} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{raddv} will be ignored.}
}
\value{
data.frame
}
\description{
Function for generating random Protocol Deviations Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID ASTDT DVTERM DVSEQ.
}
\examples{
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
raddv(ADSL, seed = 2)
}
