% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radeg.R
\name{radeg}
\alias{radeg}
\title{ECG Analysis Dataset (ADEG)}
\usage{
radeg(
  ADSL,
  egcat = c("INTERVAL", "INTERVAL", "MEASUREMENT", "FINDING"),
  param = c("QT Duration", "RR Duration", "Heart Rate", "ECG Interpretation"),
  paramcd = c("QT", "RR", "HR", "ECGINTP"),
  paramu = c("msec", "msec", "beats/min", ""),
  visit_format = "WEEK",
  n_assessments = 5L,
  n_days = 5L,
  max_n_eg = 10L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(ABLFL = c(1235, 0.1), BASE = c(NA, 0.1), BASEC = c(NA, 0.1), CHG =
    c(1234, 0.1), PCHG = c(1234, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{egcat}{(`character vector`)\cr EG category values.}

\item{param}{(`character vector`)\cr Parameter values.}

\item{paramcd}{(`character vector`)\cr Parameter code values.}

\item{paramu}{(`character vector`)\cr Parameter unit values.}

\item{visit_format}{(`character`)\cr Type of visit. Options are "WEEK" and "CYCLE".}

\item{n_assessments}{(`integer`)\cr Number of weeks or cycles.}

\item{n_days}{(`integer`)\cr Number of days within cycle (only used if `visit_format` is "CYCLE").}

\item{max_n_eg}{(`integer`)\cr Maximum number of EG results per patient. Defaults to 10.}

\item{lookup}{(`data.frame`)\cr Additional parameters.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{na_percentage}{(`proportion`)\cr Default percentage of values to be replaced by `NA`.}

\item{na_vars}{(`list`)\cr A named list where the name of each element is a column name of `ds`. Each
element of this list should be a numeric vector with two elements:
\itemize{
  \item{`seed` (`integer`)\cr The seed to be used for this element - can be `NA`.}
  \item{`percentage` (`proportion`)\cr Percentage of elements to be replaced with `NA`.
  If `NA`, `na_percentage` is used as a default.}
}}

\item{cached}{boolean whether the cached ADEG data \code{cadeg} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radeg} will be ignored.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating random dataset from ECG Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: `STUDYID`, `USUBJID`, `PARAMCD`, `BASETYPE`, `AVISITN`, `ATPTN`, `DTYPE`, `ADTM`, `EGSEQ`, `ASPID`
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)

ADEG <- radeg(ADSL, visit_format = "WEEK", n_assessments = 7L, seed = 2)
ADEG

ADEG <- radeg(ADSL, visit_format = "CYCLE", n_assessments = 2L, seed = 2)
ADEG
}
\author{
tomlinsj, npaszty, Xuefeng Hou, dipietrc
}
