% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radhy.R
\name{radhy}
\alias{radhy}
\title{Hys Law Analysis Dataset (ADHY)}
\usage{
radhy(
  ADSL,
  param = c("TBILI <= 2 times ULN and ALT value category",
    "TBILI > 2 times ULN and AST value category",
    "TBILI > 2 times ULN and ALT value category",
    "TBILI <= 2 times ULN and AST value category",
    "TBILI > 2 times ULN and ALKPH <= 2 times ULN and ALT value category",
    "TBILI > 2 times ULN and ALKPH <= 2 times ULN and AST value category",
    "TBILI > 2 times ULN and ALKPH <= 5 times ULN and ALT value category",
    "TBILI > 2 times ULN and ALKPH <= 5 times ULN and AST value category"),
  paramcd = c("BLAL", "BGAS", "BGAL", "BLAS", "BA2AL", "BA2AS", "BA5AL", "BA5AS"),
  aval_base = c("Criteria not met", ">3-5ULN", ">5-10ULN", ">10-20ULN", ">20ULN"),
  seed = NULL,
  cached = FALSE
)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{param}{As character string. list of parameter values.}

\item{paramcd}{As character string. list of parameter code values.}

\item{aval_base}{As character vector of levels for AVALC and BASEC variables.}

\item{seed}{Seed for random number generation.}

\item{cached}{boolean whether the cached ADHY data \code{cadhy} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radhy} will be ignored.}
}
\value{
data.frame
}
\description{
Function for generating random dataset from Hys Law Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID, USUBJID, PARAMCD, AVISITN, ADTM, SRCSEQ.
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
radhy(ADSL, seed = 2)
}
\author{
wojciakw
}
