% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radlb.R
\name{radlb}
\alias{radlb}
\title{Laboratory Analysis Dataset (ADLB).}
\usage{
radlb(ADSL, param = c("Alanine Aminotransferase Measurement",
  "C-Reactive Protein Measurement", "Immunoglobulin A Measurement"),
  paramcd = c("ALT", "CRP", "IGA"), visit_format = "WEEK",
  n_assessments = 5, n_days = 5, seed = NULL)
}
\arguments{
\item{ADSL}{dataset.}

\item{visit_format}{type of visit either "WEEK" or "CYCLE".}

\item{n_assessments}{number of weeks or cycles.}

\item{n_days}{number of days within cycle.}

\item{seed}{starting point used in the generation of a sequence of random numbers}

\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}
}
\value{
data.frame
}
\description{
Function for generating random dataset from Laboratory Test Findings domain for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID USUBJID PARAMCD AVISITN.
}
\examples{
ADSL <- radsl(N = 10)
ADLB <- radlb(ADSL, visit_format = "WEEK", n_assessments = 7)
ADLB <- radlb(ADSL, param = c("Immunoglobulin A Measurement", "Immunoglobulin G Measurement"), paramcd = c("IGA", "IGG"))
ADLB <- radlb(ADSL, visit_format = "CYCLE", n_assessments = 3)
head(ADLB)

}
\author{
tomlinsj, npaszty
}
