% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radmh.R
\name{radmh}
\alias{radmh}
\title{Medical History Analysis Dataset (ADMH)}
\usage{
radmh(
  ADSL,
  max_n_mhs = 10L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(MHBODSYS = c(NA, 0.1), MHDECOD = c(1234, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{max_n_mhs}{(`integer`)\cr Maximum number of MHs per patient. Defaults to 10.}

\item{lookup}{(`data.frame`)\cr Additional parameters.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{na_percentage}{(`proportion`)\cr Default percentage of values to be replaced by `NA`.}

\item{na_vars}{(`list`)\cr A named list where the name of each element is a column name of `ds`. Each
element of this list should be a numeric vector with two elements:
\itemize{
  \item{`seed` (`integer`)\cr The seed to be used for this element - can be `NA`.}
  \item{`percentage` (`proportion`)\cr Percentage of elements to be replaced with `NA`.
  If `NA`, `na_percentage` is used as a default.}
}}

\item{cached}{boolean whether the cached ADMH data \code{cadmh} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radmh} will be ignored.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating a random Medical History Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: `STUDYID`, `USUBJID`, `ASTDTM`, `MHSEQ`
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, study_duration = 2, seed = 1)

ADMH <- radmh(ADSL, seed = 2)
ADMH
}
