% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{radqlqc}
\alias{radqlqc}
\title{EORTC QLQ-C30 V3 Analysis Dataset (ADQLQC)}
\usage{
radqlqc(ADSL, percent, number, seed = NULL, cached = FALSE)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{percent}{(`numeric`)\cr Completion - Completed at least y percent of questions, 1 record per visit}

\item{number}{(`numeric`)\cr Completion - Completed at least x question(s), 1 record per visit}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{cached}{boolean whether the cached ADQLQC data \code{cadqlqc} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radqlqc} will be ignored.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating a random EORTC QLQ-C30 V3 Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, study_duration = 2, seed = 1)

ADQLQC <- radqlqc(ADSL, seed = 1, percent = 80, number = 2)
ADQLQC
}
