% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqs.R, R/radqs_original.R
\name{radqs}
\alias{radqs}
\title{Questionnaires Analysis Dataset (ADQS)}
\usage{
radqs(ADSL, seed = NULL, param = NULL, paramcd = NULL)

radqs(ADSL, seed = NULL, param = NULL, paramcd = NULL)
}
\arguments{
\item{ADSL}{dataset.}

\item{seed}{starting point used in the generation of a sequence of random numbers}

\item{param}{Vector of questionaire parameters.}

\item{paramcd}{Vector of questionaire parameter codes of the same length
as parameters and corresponding to their values.}

\item{visit_format}{type of visit either "WEEK" or "CYCLE".}

\item{n_assessments}{number of weeks or cycles.}

\item{n_days}{number of days within cycle.}

\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}

\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}

\item{param}{Vector of questionaire parameters.}

\item{paramcd}{Vector of questionaire parameter codes of the same length
as parameters and corresponding to their values.}
}
\value{
data.frame

data.frame
}
\description{
Function for generating random questionaire dataset for a given
Subject-Level Analysis Dataset.

Function for generating random questionaire dataset for a given
Subject-Level Analysis Dataset
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID USUBJID PARAMCD AVISITN.
}
\examples{
ADSL <- radsl()
ADQS <- radqs(ADSL, visit_format = "WEEK", n_assessments = 7)
ADQS <- radqs(ADSL, visit_format = "CYCLE", n_assessments = 3)
head(ADQS)

ADSL <- radsl()
ADQS <- radqs(ADSL, seed = 10)
head(ADQS)

}
\author{
npaszty
}
