% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsaftte.R
\name{radsaftte}
\alias{radsaftte}
\title{Time to Safety Event Analysis Dataset (ADSAFTTE)}
\usage{
radsaftte(ADSL, ...)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{...}{Additional argument to be passed to \code{radaette}}
}
\description{
Function to generate random time to safety event dataset for a
given subject-level analysis dataset.
}
\examples{

ADSL <- suppressWarnings(radsl(N = 10, seed = 1, study_duration = 2))
radsaftte(ADSL, seed = 2)

}
