% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsub.R
\name{radsub}
\alias{radsub}
\title{Subcategory Analysis Dataset (ADSUB)}
\usage{
radsub(
  ADSL,
  param = c("Baseline Weight", "Baseline Height", "Baseline BMI", "Baseline ECOG",
    "Baseline Biomarker Mutation"),
  paramcd = c("BWGHTSI", "BHGHTSI", "BBMISI", "BECOG", "BBMRKR1"),
  seed = NULL,
  na_percentage = 0,
  na_vars = list(),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{param}{(`character vector`)\cr Parameter values.}

\item{paramcd}{(`character vector`)\cr Parameter code values.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{na_percentage}{(`proportion`)\cr Default percentage of values to be replaced by `NA`.}

\item{na_vars}{(`list`)\cr A named list where the name of each element is a column name of `ds`. Each
element of this list should be a numeric vector with two elements:
\itemize{
  \item{`seed` (`integer`)\cr The seed to be used for this element - can be `NA`.}
  \item{`percentage` (`proportion`)\cr Percentage of elements to be replaced with `NA`.
  If `NA`, `na_percentage` is used as a default.}
}}

\item{cached}{boolean whether the cached ADSUB data \code{cadsub} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radsub} will be ignored.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating a random Subcategory Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject.

Keys: `STUDYID`, `USUBJID`, `PARAMCD`, `AVISITN`, `ADTM`, `SRCSEQ`
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)

ADSUB <- radsub(ADSL, seed = 2)
ADSUB
}
\author{
tomlinsj, npaszty, Xuefeng Hou, dipietrc
}
