% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radte.R
\name{radte}
\alias{radte}
\title{Trial Elements domain dataset (ATE)}
\usage{
radte(ADSL, seed = NULL, lookup = NULL, event.descr = NULL)
}
\arguments{
\item{ADSL}{adsl (Subject-Level Analysis Dataset) start data set}

\item{seed}{seed for random number generation}

\item{lookup}{data.frame containing additional parameters}

\item{event.descr}{character vector with description of events}
}
\value{
data.frame
}
\description{
Function for generating random Dataset from Trial Elements domain for a given
Subject-Level Analysis Dataset
}
\examples{
ADSL <- radsl()
ADTE <- radte(ADSL)
head(ADTE)

}
