% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radtr.R
\name{radtr}
\alias{radtr}
\title{Tumor Response Analysis Dataset (ADTR)}
\usage{
radtr(
  ADSL,
  param = c("Sum of Longest Diameter by Investigator"),
  paramcd = c("SLDINV"),
  seed = NULL,
  cached = FALSE,
  ...
)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{param}{(`character vector`)\cr Parameter values.}

\item{paramcd}{(`character vector`)\cr Parameter code values.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{cached}{boolean whether the cached ADTR data \code{cadtr} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radtr} will be ignored.}

\item{...}{Additional arguments to be passed to `radrs`.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating a random Tumor Response Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: `STUDYID`, `USUBJID`, `PARAMCD`, `BASETYPE`, `AVISITN`, `DTYPE`
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)

ADTR <- radtr(ADSL, seed = 2)
ADTR
}
\author{
tomlinsj, npaszty, Xuefeng Hou, dipietrc
}
