% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radvs.R
\name{radvs}
\alias{radvs}
\title{Vital Signs Dataset}
\usage{
radvs(ADSL, seed = NULL)
}
\arguments{
\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}

\item{seed}{Seed for random number generation.}
}
\value{
data.frame
}
\description{
Function for generating random dataset from laboratory Test Findings domain for a given
Subject-Level Analysis Dataset
}
\details{
One record per subject per parameter per analysis visit per analysis date.
SDTM variables are populated on new records coming from other single records.
Otherwise, SDTM variables are left blank.

Keys: STUDYID USUBJID PARAMCD AVISITN
}
\examples{

ADSL <- radsl()
ADVS <- radvs(ADSL)
head(ADVS)
}
\author{
npaszty
}
