% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{random.cdisc.data-package}
\alias{random.cdisc.data}
\alias{random.cdisc.data-package}
\title{\code{random.cdisc.data} Package}
\description{
Package to create random SDTM and ADAM datasets.
}
\author{
\strong{Maintainer}: Pawel Rucki \email{pawel.rucki@roche.com}

Authors:
\itemize{
  \item Nick Paszty \email{npaszty@gene.com}
  \item Jana Stoilova \email{jana.stoilova@roche.com}
  \item Joe Zhu \email{joe.zhu@roche.com}
  \item Christopher DiPietrantonio
  \item Adrian Waddell \email{adrian.waddell@gene.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
