% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{rasl}
\alias{rasl}
\title{Generate ASL}
\usage{
rasl(N = 400, seed = NULL)
}
\arguments{
\item{N}{number of patients}

\item{seed}{seed for random number generation}
}
\description{
Generate ASL
}
\examples{

ASL <- rasl()

}
