% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dim_control_and_checks.R
\name{reduce_num_levels_in_df}
\alias{reduce_num_levels_in_df}
\title{Reduce number of levels in a column of a df}
\usage{
reduce_num_levels_in_df(
  dt,
  variable,
  p_to_keep = 0.7,
  num_max_values = NULL,
  num_of_rare_values = 0,
  add_specific_value = NULL,
  keep_spec_rows = NULL,
  explorative = FALSE
)
}
\arguments{
\item{dt}{(\code{data.frame})\cr data.frame with too many levels for a \code{variable}.}

\item{variable}{(\code{character(1)})\cr string with the name of the column to be reduced.}

\item{p_to_keep}{(\code{numeric(1)})\cr probability cut-off to keep the most frequent levels. If \code{num_max_values} is
specified (i.e. not \code{NULL}) this is not used.}

\item{num_max_values}{(\code{integer(1)})\cr maximum number of levels to keep. This encompasses only
rare values (\code{num_of_rare_values}) but not additional desired values (\code{add_specific_value})
and rows (\code{keep_spec_rows}).}

\item{num_of_rare_values}{(\code{integer(1)})\cr number of additional rare values to keep. These are selected
from the least frequent levels.}

\item{add_specific_value}{(\code{character})\cr specific values to keep.}

\item{keep_spec_rows}{(\code{integer})\cr specific rows to keep.}

\item{explorative}{(\code{logical(1)})\cr if \code{TRUE}, a plot with the frequency distribution of levels is shown.}
}
\value{
A modified \code{data.frame} and a plot if \code{exporative = TRUE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Use this function to reduce the number of levels in a \code{data.frame} column called \code{variable}.
This function calculates the frequency distribution of levels and sets either a soft threshold
based on the probability cut-off (\code{p_to_keep}) and/or an hard threshold (\code{num_max_values}).
}
\details{
If necessary, a number of additional rare values can be picked from the least represented levels
(\code{num_of_rare_values}). For general use it is also possible to keep certain values
(\code{add_specific_value}) and rows (\code{keep_spec_rows}). Exploratory plots can be also appreciated with
\code{explorative = TRUE}.
}
\examples{
# real case scenario: trimming of variables with too many levels
#  reduce_num_levels_in_df(adae_pharmaverse, "AEDECOD",
#    num_max_values = 7, num_of_rare_values = 1,
#    add_specific_value = c(
#      "VOMITING", "NAUSEA", "SKIN IRRITATION", "HEADACHE", # For SMQ01NAM, SMQ02NAM, CQ01NAM
#      "MYOCARDIAL INFARCTION" # for aet07 AESDTH == "Y"
#    )
#  )

}
