% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_var.R, R/utils.R
\name{rel_var}
\alias{rel_var}
\title{Create a related variable within a domain}
\usage{
rel_var(df = NULL, var_name = NULL, var_values = NULL,
  related_var = NULL)

rel_var(df = NULL, var_name = NULL, var_values = NULL,
  related_var = NULL)
}
\arguments{
\item{df}{data frame containing the related variables.}

\item{var_name}{related to variable name.}

\item{var_values}{values of related to variable.}

\item{related_var}{variable name with existing values to which var_name values must relate.}

\item{visit_format}{as chara cter string. Valid values: WEEK, CYCLE}

\item{n_assessments}{number of assessments. Valid values: integer}

\item{n_days}{number of days for each cycle: Valid values: integer}

\item{n_timepoints}{number of timepoints: Valid values: integer. To Be Implemented}
}
\value{
a factor of length n_assessments
}
\description{
Create a related variable within a domain

Assign values to a related variable within a domain.
}
\examples{

rel_var(domain_name = ADLB, var_name = "PARAMCD", var_values = c("ALT", "CRP", "IGA", "IGM"), related_var = "PARAM")

rel_var(df = ADLB, var_name = "PARAMCD", var_values = c("ALT", "CRP", "IGA", "IGM"), related_var = "PARAM")

}
