% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_na}
\alias{replace_na}
\title{Replace values in a vector by NA}
\usage{
replace_na(v, percentage = 0.05, seed = NULL)
}
\arguments{
\item{v}{(\code{any}) vector of any type}

\item{percentage}{(\code{numeric}) Value between 0 and 1 defining
how much of the vector shall be replaced by NA. This number
is randomized by plus minus 5 percent to have full randomization.}

\item{seed}{(\code{numeric}) seed to set to make randomization reproducible}
}
\value{
The input vector v where a certain number of values
  is replaced by NA
}
\description{
Randomized replacement of values by NA
}
