% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retain}
\alias{retain}
\title{Primary Keys: Retain Values}
\usage{
retain(df, value_var, event, outside = NA)
}
\arguments{
\item{df}{(\code{data.frame})\cr Data frame in which to apply the retain.}

\item{value_var}{(\code{any})\cr Variable in \code{df} containing the value to be retained.}

\item{event}{(\code{expression})\cr Expression returning a logical value to trigger the retain.}

\item{outside}{(\code{any})\cr Additional value to retain. Defaults to \code{NA}.}
}
\value{
A vector of values where expression is true.
}
\description{
Retain values within primary keys.
}
\keyword{internal}
