% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rsdtm}
\alias{rsdtm}
\title{Create a Random SDTM Dataset}
\usage{
rsdtm(domain, N = 100, ..., DM = NULL, start_with = list())
}
\description{
Create a Random SDTM Dataset
}
\examples{
DM <- rsdtm("DM")
}
