% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rtexp}
\alias{rtexp}
\title{Truncated Exponential Distribution}
\usage{
rtexp(n, rate, l = NULL, r = NULL)
}
\arguments{
\item{n}{number of random numbers.}

\item{rate}{non-negative rate.}

\item{l}{positive left-hand truncation parameter.}

\item{r}{positive right-hand truncation parameter.}
}
\value{
The random numbers. If neither `l` nor `r` are provided then the usual Exponential
 distribution is used.
}
\description{
This generates random numbers from a truncated Exponential distribution,
i.e. from `X | X > l` or `X | X < r` when `X ~ Exp(rate)`. The advantage here is that
we guarantee to return exactly `n` numbers and without using a loop internally.
This can be derived from the quantile functions of the left- and right-truncated
Exponential distributions.
}
\examples{
x <- rexp(1e6, rate = 5)
x <- x[x > 0.5]
hist(x)

y <- rtexp(1e6, rate = 5, l = 0.5)
hist(y)

z <- rtexp(1e6, rate = 5, r = 0.5)
hist(z)

}
