% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rtpois}
\alias{rtpois}
\title{Zero-truncated Poisson Distribution}
\usage{
rtpois(n, lambda)
}
\arguments{
\item{n}{number of random numbers.}

\item{lambda}{non-negative mean(s).}
}
\value{
The random numbers.
}
\description{
This generates random numbers from a zero-truncated Poisson distribution,
i.e. from `X | X > 0` when `X ~ Poisson(lambda)`. The advantage here is that
we guarantee to return exactly `n` numbers and without using a loop internally.
This solution was provided in a post by
[Peter Dalgaard](https://stat.ethz.ch/pipermail/r-help/2005-May/070680.html).
}
\examples{
x <- rpois(1e6, lambda = 5)
x <- x[x > 0]
hist(x)

y <- rtpois(1e6, lambda = 5)
hist(y)

}
