% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{Example ADSL data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 18 columns.}
\usage{
data(adsl)
}
\description{
Example ADSL data
}
\examples{
data(adsl)
}
\keyword{datasets}
