% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{adsl}
\alias{adsl}
\title{Generate ADSL dataset}
\usage{
adsl(N, ..., DM = NULL, start_with = list())
}
\arguments{
\item{N}{number of pts}

\item{...}{specific ADSL variables}

\item{start_with}{named list with variable}
}
\description{
Generate ADSL dataset
}
\examples{

ASL <- radam("ASL")
head(ASL)

ASL <- radam("ASL",
   start_with = list(
      BECOG = LETTERS[1:5],
      BAGED = sample(size = nrow(ASL), letters[1:5], replace = TRUE)
   )
)
head(ASL)

}
