% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_metadata}
\alias{apply_metadata}
\title{Apply label and variable ordering attributes to domains.}
\usage{
apply_metadata(df, filename, ..., ADSL = NULL)
}
\arguments{
\item{df}{data frame to which metadata are applied.}

\item{filename}{yaml file containing domain metadata.}

\item{...}{ellipsis.}

\item{ADSL}{logical to control merging of ADSL data to domain.}
}
\description{
Apply label and variable ordering attributes to domains.
}
\examples{
\dontrun{
 ADSL <- apply_metadata(ADSL, "ADSL.yml", seed = seed, ADSL = ADSL)
 ADLB <- apply_metadata(ADLB, "ADLB.yml", seed = seed, ADSL = ADSL)
}
}
