% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_metadata}
\alias{apply_metadata}
\title{Apply label and variable ordering attributes to domains}
\usage{
apply_metadata(df, filename, ADSL = NULL)
}
\arguments{
\item{df}{data frame to which metadata are applied.}

\item{filename}{\code{yaml} file containing domain metadata.}

\item{ADSL}{logical to control merging of ADSL data to domain.}
}
\description{
Apply label and variable ordering attributes to domains
}
\examples{
seed <- 1
ADSL <- suppressWarnings(radsl(seed = seed))
ADLB <- radlb(ADSL, seed = seed)
ADSL <- random.cdisc.data:::apply_metadata(ADSL, "ADSL.yml", ADSL = ADSL)
ADLB <- random.cdisc.data:::apply_metadata(ADLB, "ADLB.yml", ADSL = ADSL)
}
