% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_metadata}
\alias{apply_metadata}
\title{Apply Metadata}
\usage{
apply_metadata(
  df,
  filename,
  add_adsl = TRUE,
  adsl_filename = "metadata/ADSL.yml"
)
}
\arguments{
\item{df}{(`data.frame`)\cr Data frame to which metadata is applied.}

\item{filename}{(`yaml`)\cr File containing domain metadata.}

\item{add_adsl}{(`logical`)\cr Should ADSL data be merged to domain.}

\item{adsl_filename}{(`yaml`)\cr File containing ADSL metadata.}
}
\description{
Apply label and variable ordering attributes to domains.
}
\examples{
seed <- 1
ADSL <- radsl(seed = seed)
ADLB <- radlb(ADSL, seed = seed)
\dontrun{
yaml_path <- file.path(path.package("random.cdisc.data"), "inst", "metadata")
ADSL <- random.cdisc.data:::apply_metadata(ADSL, file.path(yaml_path, "ADSL.yml"), FALSE)
ADLB <- random.cdisc.data:::apply_metadata(
  ADLB, file.path(yaml_path, "ADLB.yml"), TRUE,
  file.path(yaml_path, "ADSL.yml")
)
}
}
\keyword{internal}
