% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_metadata}
\alias{apply_metadata}
\title{Apply label and variable ordering attributes to domains}
\usage{
apply_metadata(df, filename, add_adsl = TRUE,
  adsl_filename = "metadata/ADSL.yml")
}
\arguments{
\item{df}{data frame to which metadata are applied.}

\item{filename}{\code{yaml} file containing domain metadata.}

\item{add_adsl}{logical to control merging of ADSL data to domain}

\item{adsl_filename}{\code{yaml} file containing ADSL metadata.}
}
\description{
Apply label and variable ordering attributes to domains
}
\examples{
seed <- 1
ADSL <- suppressWarnings(radsl(seed = seed))
ADLB <- radlb(ADSL, seed = seed)
\donttest{
ADSL <- random.cdisc.data:::apply_metadata(ADSL, "../metadata/ADSL.yml", FALSE)
ADLB <- random.cdisc.data:::apply_metadata(ADLB, "../metadata/ADLB.yml", TRUE,
  "../metadata/ADSL.yml")
}
}
