% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadab}
\alias{cadab}
\title{Cached ADAB}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58520 rows and 22 columns.
}
\usage{
data(cadab)
}
\description{
Cached ADAB data generated with \code{seed = 1}
}
\keyword{datasets}
