% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadae}
\alias{cadae}
\title{Cached ADAE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 80 columns.
}
\usage{
data(cadae)
}
\description{
Cached ADAE data generated with \code{seed = 1}
}
\keyword{datasets}
