% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadaette}
\alias{cadaette}
\title{Cached ADAETTE data generated with \code{seed = 1}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1200 rows and 39 columns.}
\usage{
data(cadaette)
}
\description{
Cached ADAETTE data generated with \code{seed = 1}
}
\keyword{datasets}
