% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadcm}
\alias{cadcm}
\title{Cached ADCM}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3685 rows and 83 columns.
}
\usage{
data(cadcm)
}
\description{
Cached ADCM data generated with \code{seed = 1}
}
\keyword{datasets}
