% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadpc}
\alias{cadpc}
\title{Cached ADPC}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18179 rows and 72 columns.
}
\usage{
data(cadpc)
}
\description{
Cached ADPC data generated with \code{seed = 1}
}
\keyword{datasets}
