% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadpp}
\alias{cadpp}
\title{Cached ADPP}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8358 rows and 53 columns.
}
\usage{
data(cadpp)
}
\description{
Cached ADPP data generated with \code{seed = 1}
}
\keyword{datasets}
