% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_data.R
\docType{data}
\name{cadtr}
\alias{cadtr}
\title{Cached ADTR}
\format{
An object of class \code{data.frame} with 2800 rows and 63 columns.
}
\usage{
data(cadtr)
}
\description{
Cached ADTR data generated with \code{seed = 1}
}
\keyword{datasets}
