% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{comp_derv}
\alias{comp_derv}
\title{Completion/Compliance Data Calculation}
\usage{
comp_derv(dataset, percent, number)
}
\arguments{
\item{dataset}{(`data.frame`)\cr dataset}

\item{percent}{(`numeric`)\cr Completion - Completed at least y percent of questions, 1 record per visit}

\item{number}{(`numeric`)\cr Completion - Completed at least x question(s), 1 record per visit}
}
\value{
`data.frame`
}
\description{
Completion/Compliance Data Calculation
}
\examples{
\dontrun{
compliance_data <- comp_derv(adqlqc, 80, 2)
}
}
\keyword{internal}
