% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{get_qs_data}
\alias{get_qs_data}
\title{Questionnaires EORTC QLQ-C30 V3.0 SDTM (QS)}
\usage{
get_qs_data(
  ADSL,
  visit_format = "CYCLE",
  n_assessments = 5L,
  n_days = 1L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(QSORRES = c(1234, 0.2), QSSTRESC = c(1234, 0.2)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{ADSL dataset}

\item{visit_format}{as character string. Valid values: WEEK, CYCLE.}

\item{n_assessments}{number of assessments. Valid values: integer.}

\item{n_days}{number of days for each cycle: Valid values: integer.}

\item{lookup}{control lookup process.}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{na_vars (\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached <%= toupper(data) %> data \code{c<%=data%>} should be
returned or new data}
}
\value{
a dataframe with SDTM questionnaire data
}
\description{
Function for generating random Questionnaires SDTM domain
}
\examples{
\dontrun{
qs <- get_qs_data(adsl, n_assessments = 5L, seed = 1, na_percentage = 0.1)
qs
}
}
\keyword{internal}
