% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{get_random_dates_between}
\alias{get_random_dates_between}
\title{Function for generating random dates between 2 dates}
\usage{
get_random_dates_between(from, to, visit_id)
}
\arguments{
\item{from}{(`datetime vector`)\cr Start date/times.}

\item{to}{(`datetime vector`)\cr End date/times.}

\item{visit_id}{(`integer vector`)\cr Visit numbers.}
}
\value{
data frame with new randomly generated dates variable.
}
\description{
Function for generating random dates between 2 dates
}
\examples{
\dontrun{
df <- df \%>\%
  group_by(USUBJID) \%>\%
  mutate(ADTM = get_random_dates_between(TRTSDTM, TRTEDTM, AVISITN))
df
}
}
\keyword{internal}
