% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsub.R
\name{h_anthropometrics_by_sex}
\alias{h_anthropometrics_by_sex}
\title{Generate anthropometric measurements for males and females.}
\usage{
h_anthropometrics_by_sex(
  df,
  seed = 1,
  id_var = "USUBJID",
  sex_var = "SEX",
  sex_var_level_male = "M",
  male_weight_in_kg = list(mean = 90.6, sd = 44.9),
  female_weight_in_kg = list(mean = 77.5, sd = 46.2),
  male_height_in_m = list(mean = 1.75, sd = 0.14),
  female_height_in_m = list(mean = 1.61, sd = 0.24)
)
}
\arguments{
\item{df}{(data frame)\cr analysis dataset.}

\item{seed}{Seed for random number generation.}

\item{id_var}{(string)\cr patient identifier variable name.}

\item{sex_var}{(string)\cr name of variable representing sex of patient.}

\item{sex_var_level_male}{(string)\cr level of \code{sex_var} representing males.}

\item{male_weight_in_kg}{(named list)\cr list of mean and sd of male weight in kilograms.}

\item{female_weight_in_kg}{(named list)\cr list of mean and sd of female weight in kilograms.}

\item{male_height_in_m}{(named list)\cr list of mean and sd of male height in metres.}

\item{female_height_in_m}{(named list)\cr list of mean and sd of female height in metres.}
}
\value{
a dataframe with anthropometric measurements for each subject in analysis dataset.
}
\description{
Anthropometric measurements are randomly generated using normal approximation.
The default mean and standard deviation values used are based on US National Health
Statistics for adults aged 20 years or over. The measurements are generated in same units
as provided to the function.
}
\details{
One record per subject.
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 5, seed = 1)

DF_with_measurements <- random.cdisc.data:::h_anthropometrics_by_sex(df = ADSL)
DF_with_measurements
}
\keyword{internal}
