% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_na}
\alias{mutate_na}
\title{Replace Values with NA}
\usage{
mutate_na(ds, na_vars = NULL, na_percentage = 0.05)
}
\arguments{
\item{ds}{(`data.frame`)\cr Any data set.}

\item{na_vars}{(`list`)\cr A named list where the name of each element is a column name of `ds`. Each
element of this list should be a numeric vector with two elements:
\itemize{
  \item{`seed` (`integer`)\cr The seed to be used for this element - can be `NA`.}
  \item{`percentage` (`proportion`)\cr Percentage of elements to be replaced with `NA`.
  If `NA`, `na_percentage` is used as a default.}
}}

\item{na_percentage}{(`proportion`)\cr Default percentage of values to be replaced by `NA`.}
}
\description{
`r lifecycle::badge("stable")`

Replace column values with `NA`s.
}
