% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_na}
\alias{mutate_na}
\title{Replace column values by NAs}
\usage{
mutate_na(ds, na_vars = NULL, na_percentage = 0.05)
}
\arguments{
\item{ds}{(\code{data.frame}) Any data set}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}
}
\description{
Replace column values by NAs
}
