% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{param_init}
\alias{param_init}
\title{Verify and initialize parameter (PARAM) values.}
\usage{
param_init(param, paramcd)
}
\arguments{
\item{param}{as character string. list of PARAM values.}

\item{paramcd}{as character string. list of PARAMCD values.}
}
\value{
a vector of n parameters
}
\description{
Verify and initialize parameter (PARAM) values.
}
\examples{
\dontrun{
param_init(c("Alanine Aminotransferase Measurement", "ALT")
}
}
