% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radab.R
\name{radab}
\alias{radab}
\title{ADA dataset}
\usage{
radab(
  ADSL,
  ADPC,
  constants = c(D = 100, ka = 0.8, ke = 1),
  paramcd = c("R1800000", "RESULT1", "R1800001", "RESULT2", "ADASTAT1", "INDUCD1",
    "ENHANC1", "TRUNAFF1", "EMERNEG1", "EMERPOS1", "PERSADA1", "TRANADA1", "TIMADA1",
    "ADADUR1", "ADASTAT2", "INDUCD2", "ENHANC2", "EMERNEG2", "EMERPOS2", "TRUNAFF2"),
  param = c("Antibody titer units", "ADA interpreted per sample result",
    "Neutralizing Antibody titer units", "NAB interpreted per sample result",
    "ADA Status of a patient", "Treatment induced ADA", "Treatment enhanced ADA",
    "Treatment unaffected ADA", "Treatment Emergent - Negative, ADA",
    "Treatment Emergent - Positive, ADA", "Persistent ADA", "Transient ADA",
    "Time to onset of ADA", "ADA Duration", "NAB Status of a patient",
    "Treatment induced ADA, Neutralizing Antibody",
    "Treatment enhanced ADA, Neutralizing Antibody", 
    
    "Treatment Emergent - Negative, Neutralizing Antibody",
    "Treatment Emergent - Positive, Neutralizing Antibody",
    "Treatment unaffected ADA, Neutralizing Antibody"),
  avalu = c("titer", "", "titer", "", "", "", "", "", "", "", "", "", "weeks", "weeks",
    "", "", "", "", "", ""),
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{ADPC}{(`character`)\cr parameter code.}

\item{constants}{(`character vector`)\cr Constant parameters to be used for creating analysis values.}

\item{paramcd}{(`character`)\cr parameter code.}

\item{param}{(`character`)\cr parameters.}

\item{avalu}{(`string`)\cr Analysis value unit value}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached ADSL data \code{cadsl} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radsl} will be ignored.}
}
\description{
ADA dataset
}
\details{
One record per study per subject per parameter per time point
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
ADPC <- radpc(ADSL, seed = 2)
ADAB <- radab(ADSL, ADPC, seed = 2)
}
