% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radae.R
\name{radae}
\alias{radae}
\title{Adverse Events Analysis Dataset}
\usage{
radae(ADSL, max_n_aes = 10, seed = NULL)
}
\arguments{
\item{ADSL}{adsl start data set}

\item{max_n_aes}{is the maximum number of adverse events per patient}

\item{seed}{for random number generation}
}
\description{
Adverse Events Analysis Dataset
}
\examples{

ADSL <- radsl()
ADAE <- radae(ADSL)

head(ADAE)

}
