% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radae.R
\name{radae}
\alias{radae}
\title{Adverse Events Analysis Dataset (ADAE)}
\usage{
radae(ADSL, max_n_aes = 10, seed = NULL, lookup = NULL)
}
\arguments{
\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}

\item{max_n_aes}{The maximum number of adverse events per patient.}

\item{seed}{Seed for random number generation.}

\item{lookup}{Data.frame containing additional parameters.}
}
\value{
data.frame
}
\description{
Function for generating random Adverse Events Analysis Dataset for a given
Subject-Level Analysis Dataset
}
\details{
One record per each record in the corresponding SDTM domain

Keys: STUDYID USUBJID ASTDTM AETERM AESEQ
}
\examples{

ADSL <- radsl()
ADAE <- radae(ADSL)

head(ADAE)

}
