% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radae.R
\name{radae}
\alias{radae}
\title{Adverse Events Analysis Dataset (ADAE).}
\usage{
radae(ADSL, max_n_aes = 10, lookup = NULL, seed = NULL)
}
\arguments{
\item{ADSL}{dataset.}

\item{max_n_aes}{maximum number of AEs per patient.}

\item{lookup}{table of AEs.}

\item{seed}{starting point used in the generation of a sequence of random numbers.}

\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}

\item{lookup}{Data.frame containing additional parameters.}
}
\value{
data.frame
}
\description{
Function for generating random Adverse Events Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID AESTDTM AETERM AESEQ.
}
\examples{
ADSL <- radsl()
ADAE <- radae(ADSL)
head(ADAE)

}
