% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radae.R
\name{radae}
\alias{radae}
\title{Adverse Event Analysis Dataset (ADAE)}
\usage{
radae(ADSL, max_n_aes = 10, lookup = NULL, seed = NULL,
  cached = FALSE)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{max_n_aes}{Maximum number of AEs per patient.}

\item{lookup}{Data.frame containing additional parameters.}

\item{seed}{Seed for random number generation.}

\item{cached}{boolean whether the cached ADAE data \code{cadae} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radae} will be ignored.}
}
\value{
data.frame
}
\description{
Function for generating random Adverse Event Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID AESTDTM AETERM AESEQ.
}
\examples{
ADSL <- radsl(seed = 1)
ADAE <- radae(ADSL, seed = 2)
head(ADAE)

}
