% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radae.R
\name{radae}
\alias{radae}
\title{Adverse Event Analysis Dataset (ADAE)}
\usage{
radae(
  ADSL,
  max_n_aes = 10L,
  lookup = NULL,
  lookup_aag = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AEBODSYS = c(NA, 0.1), AEDECOD = c(1234, 0.1), AETOXGR = c(1234, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{max_n_aes}{(`integer`)\cr Maximum number of AEs per patient. Defaults to 10.}

\item{lookup}{(`data.frame`)\cr Additional parameters.}

\item{lookup_aag}{(`data.frame`)\cr Additional metadata parameters.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{na_percentage}{(`proportion`)\cr Default percentage of values to be replaced by `NA`.}

\item{na_vars}{(`list`)\cr A named list where the name of each element is a column name of `ds`. Each
element of this list should be a numeric vector with two elements:
\itemize{
  \item{`seed` (`integer`)\cr The seed to be used for this element - can be `NA`.}
  \item{`percentage` (`proportion`)\cr Percentage of elements to be replaced with `NA`.
  If `NA`, `na_percentage` is used as a default.}
}}

\item{cached}{boolean whether the cached ADAE data \code{cadae} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radae} will be ignored.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating random Adverse Event Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: `STUDYID`, `USUBJID`, `ASTDTM`, `AETERM`, `AESEQ`
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, study_duration = 2, seed = 1)

ADAE <- radae(ADSL, seed = 2)
ADAE

# Add metadata.
AAG <- utils::read.table(
  sep = ",", header = TRUE,
  text = paste(
    "NAMVAR,SRCVAR,GRPTYPE,REFNAME,REFTERM,SCOPE",
    "CQ01NAM,AEDECOD,CUSTOM,D.2.1.5.3/A.1.1.1.1 AESI,dcd D.2.1.5.3,",
    "CQ01NAM,AEDECOD,CUSTOM,D.2.1.5.3/A.1.1.1.1 AESI,dcd A.1.1.1.1,",
    "SMQ01NAM,AEDECOD,SMQ,C.1.1.1.3/B.2.2.3.1 AESI,dcd C.1.1.1.3,BROAD",
    "SMQ01NAM,AEDECOD,SMQ,C.1.1.1.3/B.2.2.3.1 AESI,dcd B.2.2.3.1,BROAD",
    "SMQ02NAM,AEDECOD,SMQ,Y.9.9.9.9/Z.9.9.9.9 AESI,dcd Y.9.9.9.9,NARROW",
    "SMQ02NAM,AEDECOD,SMQ,Y.9.9.9.9/Z.9.9.9.9 AESI,dcd Z.9.9.9.9,NARROW",
    sep = "\n"
  ), stringsAsFactors = FALSE
)

ADAE <- radae(ADSL, lookup_aag = AAG)

with(
  ADAE,
  cbind(
    table(AEDECOD, SMQ01NAM),
    table(AEDECOD, CQ01NAM)
  )
)
}
