% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radae.R
\name{radae}
\alias{radae}
\title{Adverse Event Analysis Dataset (ADAE)}
\usage{
radae(ADSL, max_n_aes = 10L, lookup = NULL, seed = NULL,
  na_percentage = 0, na_vars = list(AEBODSYS = c(NA, 0.1), AEDECOD =
  c(1234, 0.1), AETOXGR = c(1234, 0.1)), cached = FALSE)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{max_n_aes}{Maximum number of AEs per patient.}

\item{lookup}{Data.frame containing additional parameters.}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached ADAE data \code{cadae} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radae} will be ignored.}
}
\value{
data.frame
}
\description{
Function for generating random Adverse Event Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID ASTDTM AETERM AESEQ.
}
\examples{
ADSL <- radsl(N = 10, study_duration = 2, seed = 1)
radae(ADSL, seed = 2)
}
