% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radaette.R
\name{radaette}
\alias{radaette}
\title{Time To Adverse Event Dataset}
\usage{
radaette(ADSL, seed = NULL)
}
\arguments{
\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}

\item{seed}{Seed for random number generation.}
}
\description{
Time To Adverse Event Dataset
}
\examples{
asl <- radsl()
radaette <- radaette(ADSL = asl)
head(radaette)

}
\author{
Xiuting Mi
}
