% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radaette.R
\name{radaette}
\alias{radaette}
\title{Time to Adverse Event Analysis Dataset (ADAETTE)}
\usage{
radaette(ADSL, seed = NULL, lookup = NULL, event.descr = NULL,
  censor.descr = NULL)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{seed}{Seed for random number generation.}

\item{lookup}{Data.frame containing additional parameters.}

\item{event.descr}{Character vector with description of events.}

\item{censor.descr}{Character vector with description of censors.}
}
\value{
data.frame
}
\description{
Function to generate random time to AE dataset for a
given subject-level analysis dataset.
}
\details{
Keys: STUDYID USUBJID PARAMCD AVISITN ADT TTESEQ
}
\examples{
ADSL <- radsl()
ADAETTE <- radaette(ADSL)
head(ADAETTE)

}
\author{
Xiuting Mi
}
