% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radam.R
\name{radam}
\alias{radam}
\title{Generate Dataset}
\usage{
radam(domain, N = 100, ..., ADSL = NULL, start_with = list())
}
\arguments{
\item{domain}{dataset domain}

\item{N}{number of pts}

\item{...}{arguments forwarded to data generation function}

\item{start_with}{named list with variable}

\item{ADLS}{data set (to take USUBJID from)}
}
\description{
Generate Dataset
}
\examples{

DM <- rsdtm("DM")
ASL <- radam("ASL")
radam("ATE")
radam("ARS")

}
