% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radcm.R
\name{radcm}
\alias{radcm}
\title{Concomitant Medication Analysis Dataset (ADCM)}
\usage{
radcm(ADSL, max_n_cms = 10, seed = NULL, lookup = NULL,
  cached = FALSE, random_NA = 0, NA_vars = list(CMCLAS = c(NA, 0.1),
  CMDECOD = c(1234, 0.1), ATIREL = c(1234, 0.1)))
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{max_n_cms}{maximum number of concommitant medications per patient.}

\item{seed}{Seed for random number generation.}

\item{lookup}{Data.frame containing additional parameters.}

\item{cached}{boolean whether the cached ADCM data \code{cadcm} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radcm} will be ignored.}

\item{random_NA}{(\code{numeric}) Number between 0 and 1 to define how many
NA values per data column should be generated. 1 means 100 \%}

\item{NA_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used}
}}
}
\value{
data.frame
}
\description{
Function for generating random Concomitant Medication Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID CMSTDTM CMTERM CMSEQ.
}
\examples{
ADSL <- radsl(seed = 1)
ADCM <- radcm(ADSL, seed = 2)
head(ADCM)

}
