% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radcm.R
\name{radcm}
\alias{radcm}
\title{Previous and Concomitant Medications Analysis Dataset (ADCM)}
\usage{
radcm(
  ADSL,
  max_n_cms = 10L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(CMCLAS = c(NA, 0.1), CMDECOD = c(1234, 0.1), ATIREL = c(1234, 0.1)),
  who_coding = FALSE,
  cached = FALSE
)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{max_n_cms}{(`integer`)\cr Maximum number of concomitant medications per patient. Defaults to 10.}

\item{lookup}{(`data.frame`)\cr Additional parameters.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{na_percentage}{(`proportion`)\cr Default percentage of values to be replaced by `NA`.}

\item{na_vars}{(`list`)\cr A named list where the name of each element is a column name of `ds`. Each
element of this list should be a numeric vector with two elements:
\itemize{
  \item{`seed` (`integer`)\cr The seed to be used for this element - can be `NA`.}
  \item{`percentage` (`proportion`)\cr Percentage of elements to be replaced with `NA`.
  If `NA`, `na_percentage` is used as a default.}
}}

\item{who_coding}{(`flag`)\cr Whether WHO coding (with multiple paths per medication) should be used.}

\item{cached}{boolean whether the cached ADCM data \code{cadcm} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radcm} will be ignored.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating random Concomitant Medication Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: `STUDYID`, `USUBJID`, `ASTDTM`, `CMSEQ`
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)

ADCM <- radcm(ADSL, seed = 2)
ADCM

ADCM_WHO <- radcm(ADSL, seed = 2, who_coding = TRUE)
ADCM_WHO
}
