% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radcm.R
\name{radcm}
\alias{radcm}
\title{Concomitant Medication Analysis Dataset (ADCM)}
\usage{
radcm(ADSL, max_n_cms = 10L, seed = NULL, lookup = NULL,
  cached = FALSE)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{max_n_cms}{maximum number of concommitant medications per patient.}

\item{seed}{Seed for random number generation.}

\item{lookup}{Data.frame containing additional parameters.}

\item{cached}{boolean whether the cached ADCM data \code{cadcm} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radcm} will be ignored.}
}
\value{
data.frame
}
\description{
Function for generating random Concomitant Medication Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID CMSTDTM CMTERM CMSEQ.
}
\examples{
ADSL <- radsl(seed = 1)
ADCM <- radcm(ADSL, seed = 2)
head(ADCM)
}
