% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radcm.R
\name{radcm}
\alias{radcm}
\title{Concomitant Medication Analysis Dataset (ADCM).}
\usage{
radcm(ADSL, max_n_cms = 10, seed = NULL, lookup = NULL)
}
\arguments{
\item{ADSL}{dataset.}

\item{max_n_cms}{maximum number of concommitant medications per patient.}

\item{seed}{starting point used in the generation of a sequence of random numbers.}

\item{lookup}{table of CMs.}

\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}

\item{lookup}{Data.frame containing additional parameters.}
}
\value{
data.frame
}
\description{
Function for generating random Concomitant Medication Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID CMSTDTM CMTERM CMSEQ.
}
\examples{
ADSL <- radsl()
ADCM <- radcm(ADSL)
head(ADCM)

}
