% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radcm.R
\name{radcm}
\alias{radcm}
\title{Concomitant Medication Analysis Dataset (ADCM)}
\usage{
radcm(ADSL, max_n_cms = 10, seed = NULL, lookup = NULL,
  cached = isTRUE(getOption("random.cdisc.data.cached")))
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{max_n_cms}{maximum number of concommitant medications per patient.}

\item{seed}{Seed for random number generation.}

\item{lookup}{Data.frame containing additional parameters.}

\item{cached}{Whether to use cashed data}
}
\value{
data.frame
}
\description{
Function for generating random Concomitant Medication Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID CMSTDTM CMTERM CMSEQ.
}
\examples{
ADSL <- radsl(seed = 1)
ADCM <- radcm(ADSL, seed = 2)
head(ADCM)

}
