% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radex.R
\name{radex}
\alias{radex}
\title{Exposure Analysis Dataset (ADEX)}
\usage{
radex(
  ADSL,
  param = c("Dose administered during constant dosing interval",
    "Number of doses administered during constant dosing interval",
    "Total dose administered", "Total number of doses administered"),
  paramcd = c("DOSE", "NDOSE", "TDOSE", "TNDOSE"),
  paramu = c("mg", " ", "mg", " "),
  parcat1 = c("INDIVIDUAL", "OVERALL"),
  parcat2 = c("Drug A", "Drug B"),
  visit_format = "WEEK",
  n_assessments = 5L,
  n_days = 5L,
  max_n_exs = 6L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1), AVALU = c(NA), 0.1),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{param}{As character string. list of parameter values.}

\item{paramcd}{As character string. list of parameter code values.}

\item{paramu}{As character vector of parameter unit value}

\item{parcat1}{Individual/Overall dose amount}

\item{parcat2}{Type of drug received Drug A / Drug B}

\item{visit_format}{Type of visit either "WEEK" or "CYCLE".}

\item{n_assessments}{Number of weeks or cycles.}

\item{n_days}{Number of days within cycle.}

\item{max_n_exs}{As numeric. maximum number of exposures}

\item{lookup}{control lookup process}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached ADEX data \code{cadex} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radex} will be ignored.}
}
\value{
data.frame
}
\description{
Function for generating random Exposure Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID EXSEQ PARAMCD PARCAT1 ASTDTM AVISITN
}
\examples{
library(random.cdisc.data)
library(utils.nest)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADEX <- radex(cached = TRUE)

}
