% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radex.R
\name{radex}
\alias{radex}
\title{Exposure Analysis Dataset (ADEX)}
\usage{
radex(
  ADSL,
  param = c("Dose administered during constant dosing interval",
    "Number of doses administered during constant dosing interval",
    "Total dose administered", "Total number of doses administered"),
  paramcd = c("DOSE", "NDOSE", "TDOSE", "TNDOSE"),
  paramu = c("mg", " ", "mg", " "),
  parcat1 = c("INDIVIDUAL", "OVERALL"),
  parcat2 = c("Drug A", "Drug B"),
  visit_format = "WEEK",
  n_assessments = 5L,
  n_days = 5L,
  max_n_exs = 6L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1), AVALU = c(NA), 0.1),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{param}{(`character vector`)\cr Parameter values.}

\item{paramcd}{(`character vector`)\cr Parameter code values.}

\item{paramu}{(`character vector`)\cr Parameter unit values.}

\item{parcat1}{(`character vector`)\cr Dose amount categories. Defaults to "Individual" and "Overall".}

\item{parcat2}{(`character vector`)\cr Types of drug received. Defaults to "Drug A" and "Drug B".}

\item{visit_format}{(`character`)\cr Type of visit. Options are "WEEK" and "CYCLE".}

\item{n_assessments}{(`integer`)\cr Number of weeks or cycles.}

\item{n_days}{(`integer`)\cr Number of days within cycle (only used if `visit_format` is "CYCLE").}

\item{max_n_exs}{(`integer`)\cr Maximum number of exposures per patient. Defaults to 6.}

\item{lookup}{(`data.frame`)\cr Additional parameters.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{na_percentage}{(`proportion`)\cr Default percentage of values to be replaced by `NA`.}

\item{na_vars}{(`list`)\cr A named list where the name of each element is a column name of `ds`. Each
element of this list should be a numeric vector with two elements:
\itemize{
  \item{`seed` (`integer`)\cr The seed to be used for this element - can be `NA`.}
  \item{`percentage` (`proportion`)\cr Percentage of elements to be replaced with `NA`.
  If `NA`, `na_percentage` is used as a default.}
}}

\item{cached}{boolean whether the cached ADEX data \code{cadex} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radex} will be ignored.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating random Exposure Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per each record in the corresponding SDTM domain.

Keys: `STUDYID`, `USUBJID`, `EXSEQ`, `PARAMCD`, `PARCAT1`, `ASTDTM`, `AENDTM`, `ASTDY`, `AENDY`,
`AVISITN`, `EXDOSFRQ`, `EXROUTE`, `VISIT`, `VISITDY`, `EXSTDTC`, `EXENDTC`, `EXSTDY`, `EXENDY`
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, study_duration = 2, seed = 1)

ADEX <- radex(ADSL, seed = 2)
ADEX
}
