% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radlb.R
\name{radlb}
\alias{radlb}
\title{Laboratory Data Analysis Dataset (ADLB)}
\usage{
radlb(ADSL, param = c("Alanine Aminotransferase Measurement",
  "C-Reactive Protein Measurement", "Immunoglobulin A Measurement"),
  paramcd = c("ALT", "CRP", "IGA"), paramu = c("U/L", "mg/L", "g/L"),
  visit_format = "WEEK", n_assessments = 5, n_days = 5,
  seed = NULL, cached = FALSE, random_NA = 0, NA_vars = list(LOQFL
  = c(NA, 0.1), ABLFL2 = c(1234, 0.1), ABLFL = c(1235, 0.1), BASE2 = c(NA,
  0.1), BASE = c(NA, 0.1), CHG2 = c(1235, 0.1), PCHG2 = c(1235, 0.1), CHG =
  c(1234, 0.1), PCHG = c(1234, 0.1)))
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{param}{As character string. list of parameter values.}

\item{paramcd}{As character string. list of parameter code values.}

\item{paramu}{As character string. list of parameter unit values.}

\item{visit_format}{Type of visit either "WEEK" or "CYCLE".}

\item{n_assessments}{Number of weeks or cycles.}

\item{n_days}{Number of days within cycle.}

\item{seed}{Seed for random number generation.}

\item{cached}{boolean whether the cached ADLB data \code{cadlb} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radlb} will be ignored.}

\item{random_NA}{(\code{numeric}) Number between 0 and 1 to define how many
NA values per data column should be generated. 1 means 100 \%}

\item{NA_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used}
}}
}
\value{
data.frame
}
\description{
Function for generating random dataset from Laboratory Data Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID USUBJID PARAMCD AVISITN.
}
\examples{
ADSL <- radsl(N = 10, seed = 1)
ADLB <- radlb(ADSL, visit_format = "WEEK", n_assessments = 7, seed = 2)
ADLB <- radlb(ADSL, param = c("Immunoglobulin A Measurement", "Immunoglobulin G Measurement"),
paramcd = c("IGA", "IGG"), paramu = c("g/L", "g/L"), seed = 2)
ADLB <- radlb(ADSL, visit_format = "CYCLE", n_assessments = 3, seed = 2)
head(ADLB)

}
\author{
tomlinsj, npaszty
}
