% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radlb.R
\name{radlb}
\alias{radlb}
\title{Laboratory Data Analysis Dataset (ADLB)}
\usage{
radlb(
  ADSL,
  lbcat = c("CHEMISTRY", "CHEMISTRY", "IMMUNOLOGY"),
  param = c("Alanine Aminotransferase Measurement", "C-Reactive Protein Measurement",
    "Immunoglobulin A Measurement"),
  paramcd = c("ALT", "CRP", "IGA"),
  paramu = c("U/L", "mg/L", "g/L"),
  aval_mean = c(18, 9, 2.2),
  visit_format = "WEEK",
  n_assessments = 5L,
  n_days = 5L,
  max_n_lbs = 10L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(LOQFL = c(NA, 0.1), ABLFL2 = c(1234, 0.1), ABLFL = c(1235, 0.1), BASE2 =
    c(NA, 0.1), BASE = c(NA, 0.1), CHG2 = c(1235, 0.1), PCHG2 = c(1235, 0.1), CHG =
    c(1234, 0.1), PCHG = c(1234, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{lbcat}{(`character vector`)\cr LB category values.}

\item{param}{(`character vector`)\cr Parameter values.}

\item{paramcd}{(`character vector`)\cr Parameter code values.}

\item{paramu}{(`character vector`)\cr Parameter unit values.}

\item{aval_mean}{(`numeric vector`)\cr Mean values corresponding to each parameter.}

\item{visit_format}{(`character`)\cr Type of visit. Options are "WEEK" and "CYCLE".}

\item{n_assessments}{(`integer`)\cr Number of weeks or cycles.}

\item{n_days}{(`integer`)\cr Number of days within cycle (only used if `visit_format` is "CYCLE").}

\item{max_n_lbs}{(`integer`)\cr Maximum number of labs per patient. Defaults to 10.}

\item{lookup}{(`data.frame`)\cr Additional parameters.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{na_percentage}{(`proportion`)\cr Default percentage of values to be replaced by `NA`.}

\item{na_vars}{(`list`)\cr A named list where the name of each element is a column name of `ds`. Each
element of this list should be a numeric vector with two elements:
\itemize{
  \item{`seed` (`integer`)\cr The seed to be used for this element - can be `NA`.}
  \item{`percentage` (`proportion`)\cr Percentage of elements to be replaced with `NA`.
  If `NA`, `na_percentage` is used as a default.}
}}

\item{cached}{boolean whether the cached ADLB data \code{cadlb} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radlb} will be ignored.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating a random Laboratory Data Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: `STUDYID`, `USUBJID`, `PARAMCD`, `BASETYPE`, `AVISITN`, `ATPTN`, `DTYPE`, `ADTM`, `LBSEQ`, `ASPID`
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)

ADLB <- radlb(ADSL, visit_format = "WEEK", n_assessments = 7L, seed = 2)
ADLB

ADLB <- radlb(ADSL, visit_format = "CYCLE", n_assessments = 2L, seed = 2)
ADLB
}
\author{
tomlinsj, npaszty, Xuefeng Hou
}
