% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radlb.R
\name{radlb}
\alias{radlb}
\title{Laboratory Data Analysis Dataset (ADLB)}
\usage{
radlb(ADSL, param = c("Alanine Aminotransferase Measurement",
  "C-Reactive Protein Measurement", "Immunoglobulin A Measurement"),
  paramcd = c("ALT", "CRP", "IGA"), paramu = c("U/L", "mg/L", "g/L"),
  visit_format = "WEEK", n_assessments = 5, n_days = 5,
  seed = NULL)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{param}{As character string. list of parameter values.}

\item{paramcd}{As character string. list of parameter code values.}

\item{paramu}{As character string. list of parameter unit values.}

\item{visit_format}{Type of visit either "WEEK" or "CYCLE".}

\item{n_assessments}{Number of weeks or cycles.}

\item{n_days}{Number of days within cycle.}

\item{seed}{Seed for random number generation.}
}
\value{
data.frame
}
\description{
Function for generating random dataset from Laboratory Data Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID USUBJID PARAMCD AVISITN.
}
\examples{
ADSL <- radsl(N = 10, seed = 1)
ADLB <- radlb(ADSL, visit_format = "WEEK", n_assessments = 7, seed = 2)
ADLB <- radlb(ADSL, param = c("Immunoglobulin A Measurement", "Immunoglobulin G Measurement"),
paramcd = c("IGA", "IGG"), paramu = c("g/L", "g/L"), seed = 2)
ADLB <- radlb(ADSL, visit_format = "CYCLE", n_assessments = 3, seed = 2)
head(ADLB)

}
\author{
tomlinsj, npaszty
}
