% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radpc.R
\name{radpc}
\alias{radpc}
\title{Title}
\usage{
radpc(
  ADSL,
  avalu = "ug/mL",
  base_mean = 50,
  rel_increase_by_visit = 0.5,
  rel_increase_by_plasma = 0.5
)
}
\arguments{
\item{rel_increase_by_visit}{}
}
\value{

}
\description{
Title
}
\details{
One record per per study per subject per parameter per time point
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
ADPC <- radpc(ADSL)
}
