% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radpc.R
\name{radpc}
\alias{radpc}
\title{Pharmacokinetics Analysis Dataset (ADPC)}
\usage{
radpc(
  ADSL,
  avalu = "ug/mL",
  constants = c(D = 100, ka = 0.8, ke = 1),
  duration = 2,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{avalu}{(`character`)\cr Analysis value units.}

\item{constants}{(`character vector`)\cr Constant parameters to be used in formulas for creating analysis values.}

\item{duration}{(`numeric`)\cr Duration in number of days.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}

\item{na_percentage}{(`proportion`)\cr Default percentage of values to be replaced by `NA`.}

\item{na_vars}{(`list`)\cr A named list where the name of each element is a column name of `ds`. Each
element of this list should be a numeric vector with two elements:
\itemize{
  \item{`seed` (`integer`)\cr The seed to be used for this element - can be `NA`.}
  \item{`percentage` (`proportion`)\cr Percentage of elements to be replaced with `NA`.
  If `NA`, `na_percentage` is used as a default.}
}}

\item{cached}{boolean whether the cached ADPC data \code{cadpc} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radpc} will be ignored.}
}
\value{
`data.frame`
}
\description{
`r lifecycle::badge("stable")`

Function for generating a random Pharmacokinetics Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per per study per subject per parameter per time point.
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)

ADPC <- radpc(ADSL, seed = 2)
ADPC

ADPC <- radpc(ADSL, seed = 2, duration = 3)
ADPC
}
