% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radpc.R
\name{radpc}
\alias{radpc}
\title{Pharmacokinetics Analysis Dataset (ADPC)}
\usage{
radpc(
  ADSL,
  avalu = "ug/mL",
  constants = c(D = 100, ka = 0.8, ke = 1),
  duration = 2,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{avalu}{(`string`)\cr Analysis value unit value}

\item{constants}{(`character vector`)\cr Constant parameters to be used in the PK
equation for creating analysis values.}

\item{duration}{(`numeric`)\cr Duration in number of days.}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached ADSL data \code{cadsl} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radsl} will be ignored.}
}
\description{
Pharmacokinetics Analysis Dataset (ADPC)
}
\details{
One record per per study per subject per parameter per time point
}
\examples{
library(random.cdisc.data)
adsl <- radsl(N = 10, seed = 1, study_duration = 2)

adpc <- radpc(adsl, seed = 2)
adpc

adpc <- radpc(adsl, seed = 2, duration = 3)
adpc
}
