% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radpc.R
\name{radpc}
\alias{radpc}
\title{PK dataset}
\usage{
radpc(
  ADSL,
  avalu = "ug/mL",
  base_mean = 50,
  rel_increase_by_visit = 0.5,
  rel_increase_by_plasma = 0.5,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{avalu}{(`string`)\cr Analysis value unit value}

\item{base_mean}{(`numeric`)}

\item{rel_increase_by_visit}{(`numeric`)\cr relative increase in analysis value over visits.}

\item{rel_increase_by_plasma}{(`numeric`)\cr relative increase in analysis value between parameters.}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached ADSL data \code{cadsl} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radsl} will be ignored.}
}
\value{

}
\description{
PK dataset
}
\details{
One record per per study per subject per parameter per time point
}
\examples{
library(random.cdisc.data)
library(dplyr)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
ADPC <- radpc(ADSL, seed = 2)

}
