% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radpc.R, R/radpp.R
\name{radpc}
\alias{radpc}
\title{PK dataset}
\usage{
radpc(
  ADSL,
  ppcat = c("Plasma Drug X", "Plasma Drug Y"),
  ppspec = c("Plasma"),
  paramcd = c("AUCIFO", "CMAX", "CLO"),
  param = c("AUC Infinity Obs", "Max Conc", "Total CL Obs"),
  paramu = c("day*ug/mL", "ug/mL", "ml/day/kg"),
  aval_mean = c(200, 30, 5),
  visit_format = "CYCLE",
  n_days = 2L,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1)),
  cached = FALSE
)

radpc(
  ADSL,
  ppcat = c("Plasma Drug X", "Plasma Drug Y"),
  ppspec = c("Plasma"),
  paramcd = c("AUCIFO", "CMAX", "CLO"),
  param = c("AUC Infinity Obs", "Max Conc", "Total CL Obs"),
  paramu = c("day*ug/mL", "ug/mL", "ml/day/kg"),
  aval_mean = c(200, 30, 5),
  visit_format = "CYCLE",
  n_days = 2L,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{ppcat}{(`character`)\cr category of parameter.}

\item{paramcd}{(`character`)\cr parameter code.}

\item{paramu}{(`string`)\cr Analysis value unit value.}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached ADSL data \code{cadsl} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radsl} will be ignored.}

\item{avalu}{(`string`)\cr Analysis value unit value}

\item{constants}{(`character vector`)\cr Constant parameters to be used in the PK
equation for creating analysis values.}

\item{ppsec}{(`string`)\cr spec of parameter.}

\item{param(`character`)\cr}{parameters.}

\item{aval_mean(`vector`)\cr}{the mean for corresponding parameter.}
}
\description{
PK dataset

PK parameters dataset
}
\details{
One record per per study per subject per parameter per time point

One record per per study per subject per parameter category per parameter per visit.
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
ADPC <- radpc(ADSL, seed = 2)

library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
ADPP <- radpp(ADSL, seed = 2)

}
