% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radpp.R
\name{radpp}
\alias{radpp}
\title{PK parameters dataset}
\usage{
radpp(
  ADSL,
  ppcat = c("Plasma Drug X", "Plasma Drug Y"),
  ppspec = c("Plasma", "Plasma", "Plasma", "Matrix of PD", "Matrix of PD", "Urine",
    "Urine", "Urine", "Urine"),
  paramcd = c("AUCIFO", "CMAX", "CLO", "RMAX", "TON", "RENALCL", "RENALCLD", "RCAMINT",
    "RCPCINT"),
  param = c("AUC Infinity Obs", "Max Conc", "Total CL Obs", "Time of Maximum Response",
    "Time to Onset", "Renal CL", "Renal CL Norm by Dose", "Amt Rec from T1 to T2",
    "Pct Rec from T1 to T2"),
  paramu = c("day*ug/mL", "ug/mL", "ml/day/kg", "hr", "hr", "L/hr", "L/hr/mg", "mg",
    "\%"),
  aval_mean = c(200, 30, 5, 10, 3, 0.05, 0.005, 1.5613, 15.65),
  visit_format = "CYCLE",
  n_days = 2L,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1)),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{ppcat}{(`character`)\cr category of parameter.}

\item{ppspec}{(`string`)\cr specimen material type.}

\item{paramcd}{(`character`)\cr parameter code.}

\item{param}{(`character`)\cr parameters.}

\item{paramu}{(`character`)\cr Analysis value unit value.}

\item{aval_mean}{(`vector`)\cr the mean for corresponding parameter.}

\item{visit_format}{(`string`)\cr the visit format to produce AVISIT in the dataset.}

\item{n_days}{(`integer`)\cr the number days that will be included to produce AVISIT in the dataset.}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached ADSL data \code{cadsl} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radsl} will be ignored.}
}
\description{
PK parameters dataset
}
\details{
One record per per study per subject per parameter category per parameter per visit.
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
ADPP <- radpp(ADSL, seed = 2)
}
