% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqlqc.R
\name{radqlqc}
\alias{radqlqc}
\title{EORTC QLQ-C30 V3 Analysis Dataset (ADQLQC)}
\usage{
radqlqc(ADSL, percent, number, cached = FALSE)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset}

\item{percent}{Completion - Completed at least y percent of questions, 1 record per visit}

\item{number}{Completion - Completed at least x question(s), 1 record per visit}

\item{cached}{boolean whether the cached <%= toupper(data) %> data \code{c<%=data%>} should be
returned or new data}
}
\value{
a dataframe with EORTC QLQ-C30 V3 Analysis Dataset
}
\description{
EORTC QLQ-C30 V3 Analysis Dataset (ADQLQC)
}
\examples{
\dontrun{
ADSL <- radsl(N = 10, seed = 1, study_duration = 1)
ADQLQC <- radqlqc(ADSL, percent = 80, number = 2)
}
}
