% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqs.R
\name{radqs}
\alias{radqs}
\title{Generate a random Questionaires Dataset}
\usage{
radqs(ADSL, seed = NULL)
}
\arguments{
\item{ADSL}{adsl (Subject-Level Analysis Dataset) start data set}

\item{seed}{for random number generation}
}
\description{
Function for generating random questionaire dataset for a given
Subject-Level Analysis Dataset
}
\examples{
ADSL <- radsl()
ADQS <- radqs(ADSL)
head(ADQS)

}
