% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqs.R
\name{radqs}
\alias{radqs}
\title{Generate a random Questionaires Dataset}
\usage{
radqs(ADSL, seed = NULL)
}
\arguments{
\item{ADSL}{adsl (Subject-Level Analysis Dataset) start data set}

\item{seed}{for random number generation}
}
\value{
data frame containing generated random Questionaires dataset for
Subject-Level Analysis Dataset. The dataset consists of following variables:\cr
[,1] USUBJID (Unique Subject Identifier),\cr
[,2] STUDYID (Study Identifier),\cr
[,3] SITEID (Study Site Identifier),\cr
[,4] PARAMCD (Parameter Code), \cr
[,5] PARAM (Parameter Description),\cr
[,6] AVAL (Analysis Value),\cr
[,7] AVISIT (Analysis Visit), \cr
[,8] AVISITN (Analysis Visit (N)),\cr
[,9] ABLFL (Baseline Record Flag),\cr
[,10] APBFL,\cr
[,11] CHG (Change From Baseline),\cr
[,12] PCHG (Percent Change From Baseline),\cr
}
\description{
Function for generating random questionaire dataset for a given
Subject-Level Analysis Dataset
}
\examples{
ADSL <- radsl()
ADQS <- radqs(ADSL)
head(ADQS)

}
