% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqs.R
\name{radqs}
\alias{radqs}
\title{Questionaires Dataset}
\usage{
radqs(ADSL, seed = NULL, param = NULL, paramcd = NULL)
}
\arguments{
\item{ADSL}{ADSL (Subject-Level Analysis Dataset) start data set.}

\item{seed}{Seed for random number generation.}

\item{param}{Vector of questionaire parameters.}

\item{paramcd}{Vector of questionaire parameter codes of the same length
as parameters and corresponding to their values.}
}
\value{
data.frame
}
\description{
Function for generating random questionaire dataset for a given
Subject-Level Analysis Dataset
}
\examples{
ADSL <- radsl()
ADQS <- radqs(ADSL)
head(ADQS)

}
