% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radqs.R
\name{radqs}
\alias{radqs}
\title{Questionnaires Analysis Dataset (ADQS)}
\usage{
radqs(ADSL, param = c("BFI All Questions", "Fatigue Interference",
  "Function/Well-Being (GF1,GF3,GF7)",
  "Treatment Side Effects (GP2,C5,GP5)", "FKSI-19 All Questions"),
  paramcd = c("BFIALL", "FATIGI", "FKSI-FWB", "FKSI-TSE", "FKSIALL"),
  visit_format = "WEEK", n_assessments = 5, n_days = 5,
  seed = NULL, cached = FALSE, NA_percentage = 0,
  NA_vars = list(LOQFL = c(NA, 0.1), ABLFL2 = c(1234, 0.1), ABLFL =
  c(1235, 0.1), CHG2 = c(1235, 0.1), PCHG2 = c(1235, 0.1), CHG = c(1234,
  0.1), PCHG = c(1234, 0.1)))
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{param}{As character string. list of parameter values.}

\item{paramcd}{As character string. list of parameter code values.}

\item{visit_format}{Type of visit either "WEEK" or "CYCLE".}

\item{n_assessments}{Number of weeks or cycles.}

\item{n_days}{Number of days within cycle.}

\item{seed}{Seed for random number generation.}

\item{cached}{boolean whether the cached ADQS data \code{cadqs} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radqs} will be ignored.}

\item{NA_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{NA_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}
}
\value{
data.frame
}
\description{
Function for generating random Questionaires Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID USUBJID PARAMCD AVISITN.
}
\examples{
ADSL <- radsl(seed = 1)
ADQS <- radqs(ADSL, visit_format = "WEEK", n_assessments = 7, seed = 2)
ADQS <- radqs(ADSL, visit_format = "CYCLE", n_assessments = 3, seed = 2)
head(ADQS)

}
\author{
npaszty
}
