% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radrs.R
\name{radrs}
\alias{radrs}
\title{Tumor Response Analysis Data Set (ADRS)}
\usage{
radrs(ADSL, seed = NULL, avalc = NULL, lookup = NULL)
}
\arguments{
\item{ADSL}{adsl (Subject-Level Analysis Dataset) start data set}

\item{seed}{seed for random number generation}

\item{avalc}{vector of analysis value categories}

\item{lookup}{data.frame containing additional parameters}
}
\value{
data.frame
}
\description{
Tumor Response Analysis Data Set
}
\details{
One record per subject per parameter per analysis visit per analysis date.
SDTM variables are populated on new records coming from other single records.
Otherwise, SDTM variables are left blank.

Keys: STUDYID USUBJID PARAMCD AVISITN ADT RSSEQ
}
\examples{
 ADSL <- radsl()
 ADRS <- radrs(ADSL)
 head(ADRS)
}
