% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radrs.R
\name{radrs}
\alias{radrs}
\title{Random Response Analysis Dataset}
\usage{
radrs(ADSL, seed = NULL)
}
\arguments{
\item{ADSL}{adsl (Subject-Level Analysis Dataset) start data set}

\item{seed}{for random number generation}
}
\value{
a data frame containing generated Response Analysis Dataset for
Subject-Level Analysis Dataset. The dataset consists of following variables:
STUDYID (Study Identifier), SITEID (Site Identifier), USUBJID (Unique Subject
Identifier), PARAMCD (Parameter Code), PARAM (Parameter Name),
AVALC (Analysis Value C), AVAL (Analysis Value), AVISIT (Analysis Visit)
}
\description{
Function for generating random Response Analysis Dataset for a given
Subject-Level Analysis Dataset
}
\examples{
ADSL <- radsl()
ADRS <- radrs(ADSL)
head(ADRS)
}
