% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radrs.R
\name{radrs}
\alias{radrs}
\title{Tumor Response Analysis Data Set (ADRS)}
\usage{
radrs(ADSL, seed = NULL, avalc = NULL, lookup = NULL,
  cached = isTRUE(getOption("random.cdisc.data.cached")))
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{seed}{Seed for random number generation.}

\item{avalc}{Vector of analysis value categories.}

\item{lookup}{Data.frame containing additional parameters.}

\item{cached}{Whether to use cashed data}
}
\value{
data.frame
}
\description{
Function for generating random Tumor Response Analysis Data Set for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.
SDTM variables are populated on new records coming from other single records.
Otherwise, SDTM variables are left blank.

Keys: STUDYID USUBJID PARAMCD AVISITN ADT RSSEQ
}
\examples{
 ADSL <- radsl(seed = 1)
 ADRS <- radrs(ADSL, seed = 2)
 head(ADRS)
}
