% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsaftte.R
\name{radsaftte}
\alias{radsaftte}
\title{Time to Safety Event Analysis Dataset (ADSAFTTE)}
\usage{
radsaftte(ADSL, ...)
}
\arguments{
\item{ADSL}{(`data.frame`)\cr Subject-Level Analysis Dataset (ADSL).}

\item{...}{Additional arguments to be passed to `radaette`}

\item{cached}{boolean whether the cached ADSAFTTE data \code{cadsaftte} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radsaftte} will be ignored.}
}
\value{
`data.frame`
}
\description{
Function to generate random Time-to-Safety Event Dataset for a
given Subject-Level Analysis Dataset.
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)

ADSAFTTE <- radsaftte(ADSL, seed = 2)
ADSAFTTE
}
