% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsl.R
\name{radsl}
\alias{radsl}
\title{Generate a random ADSL dataset}
\usage{
radsl(N = 400, seed = NULL)
}
\arguments{
\item{N}{number of patients}

\item{seed}{seed for random number generation}
}
\description{
ADSL is the subjectlevel dataset
}
\examples{

ADSL <- radsl()

}
