% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsl.R
\name{radsl}
\alias{radsl}
\title{Subject-Level Analysis Dataset (ADSL)}
\usage{
radsl(N = 400, seed = NULL)
}
\arguments{
\item{N}{Number of patients.}

\item{seed}{Seed for random number generation.}
}
\value{
data.frame
}
\description{
The Subject-Level Analysis Data set (ADSL) is used to provide the variables
that describe attributes of a subject. ADSL is a source for subject-level
variables used in other analysis data sets, such as population flags and
treatment variables. There is only one ADSL per study. ADSL and its related
metadata are required in a CDISC-based submission of data from a clinical
trial even if no other analysis data sets are submitted.
}
\details{
One record per subject.

Keys: STUDYID USUBJID
}
\examples{
ADSL <- radsl(N = 10)
head(ADSL)

}
