% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsl.R
\name{radsl}
\alias{radsl}
\title{Subject-Level Analysis Dataset (ADSL)}
\usage{
radsl(N = 400, seed = NULL, cached = FALSE, random_NA = 0,
  NA_vars = list(AGE = NA, SEX = NA, RACE = NA, STRATA1 = NA, STRATA2 =
  NA, BMRKR1 = c(seed = 1234, percentage = 0.1), BMRKR2 = c(1234, 0.1),
  BEP01FL = NA))
}
\arguments{
\item{N}{Number of patients.}

\item{seed}{Seed for random number generation.}

\item{cached}{boolean whether the cached ADSL data \code{cadsl} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radsl} will be ignored.}

\item{random_NA}{(\code{numeric}) Number between 0 and 1 to define how many
NA values per data column should be generated. 1 means 100 \%}

\item{NA_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used}
}}
}
\value{
data.frame
}
\description{
The Subject-Level Analysis Data set (ADSL) is used to provide the variables
that describe attributes of a subject. ADSL is a source for subject-level
variables used in other analysis data sets, such as population flags and
treatment variables. There is only one ADSL per study. ADSL and its related
metadata are required in a CDISC-based submission of data from a clinical
trial even if no other analysis data sets are submitted.
}
\details{
One record per subject.

Keys: STUDYID USUBJID
}
\examples{
ADSL <- radsl(N = 10, seed = 1)
head(ADSL)

attr(ADSL, "label") # or also tern::label(ADSL)

ADSL <- radsl(N = 20, random_NA = 0.1, seed = 1)
print(which(is.na(ADSL$SEX)))
print(which(is.na(ADSL$AGE)))

ADSL <- radsl(N = 20, random_NA = 0.1, NA_vars = list(SEX = 2, AGE = 2), seed = 1)
print(which(is.na(ADSL$SEX)))
print(which(is.na(ADSL$AGE)))

}
