% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsub.R
\name{radsub}
\alias{radsub}
\title{Subcategory Analysis Dataset (ADSUB)}
\usage{
radsub(
  ADSL,
  param = c("Baseline Weight", "Baseline Height", "Baseline BMI", "Baseline ECOG",
    "Baseline Biomarker Mutation"),
  paramcd = c("BWGHTSI", "BHGHTSI", "BBMISI", "BECOG", "BBMRKR1"),
  seed = NULL,
  na_percentage = 0,
  na_vars = list(),
  cached = FALSE
)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{param}{param character vector with parameter value strings.}

\item{paramcd}{paramcd character vector with parameter code values.}

\item{seed}{Seed for random number generation.}

\item{na_percentage}{(\code{numeric}) Default percentage of values to be replaced by NA}

\item{na_vars}{(\code{list}) A named list where the name of each element is a column name of \code{ds}. Each
 element of this list should be a numeric vector with two elements
 \itemize{
    \item{seed }{The seed to be used for this element - can be left NA}
    \item{percentage }{How many element should be replaced. 0 is 0 \% 1 is 100 \%, can be left NA and default
    percentage is used. This will overwrite default percentage (percentage argument))}
}}

\item{cached}{boolean whether the cached ADSUB data \code{cadsub} should be returned or new data
should be generated. If set to \code{TRUE} then the other arguments to \code{radsub} will be ignored.}
}
\value{
data.frame
}
\description{
Function for generating random dataset from Subcategory Analysis Dataset for a given
Subject-Level Analysis Dataset
}
\details{
One record per subject

Keys: STUDYID, USUBJID, PARAMCD, AVISITN, ADTM, SRCSEQ
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)

ADSUB <- radsub(ADSL, seed = 2)
ADSUB
}
\author{
tomlinsj, npaszty, Xuefeng Hou, dipietrc
}
