% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radtte.R
\name{radtte}
\alias{radtte}
\title{Time-to-Event Analysis dataset (ATE)}
\usage{
radtte(ADSL, lookup = NULL, event.descr = NULL, seed = NULL)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{lookup}{Data.frame containing additional parameters.}

\item{event.descr}{Character vector with description of events.

Keys: STUDYID USUBJID PARAMCD AVISITN ADT TTESEQ}

\item{seed}{Seed for random number generation.}
}
\value{
data.frame
}
\description{
Function for generating random Time-to-Event Analysis Dataset for a given
Subject-Level Analysis Dataset
}
\examples{
ADSL <- radsl()
ADTTE <- radtte(ADSL)
head(ADTTE)

}
