% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radvs.R
\name{radvs}
\alias{radvs}
\title{Vital Signs Analysis Dataset (ADVS)}
\usage{
radvs(ADSL, param = c("Diastolic Blood Pressure", "Pulse Rate",
  "Respiratory Rate", "Systolic Blood Pressure", "Temperature", "Weight"),
  paramcd = c("DIABP", "PULSE", "RESP", "SYSBP", "TEMP", "WEIGHT"),
  paramu = c("Pa", "beats/min", "breaths/min", "Pa", "C", "Kg"),
  visit_format = "WEEK", n_assessments = 5, n_days = 5,
  seed = NULL)
}
\arguments{
\item{ADSL}{Subject-Level Analysis Dataset (ADSL).}

\item{param}{As character string. list of parameter values.}

\item{paramcd}{As character string. list of parameter code values.}

\item{paramu}{As character string. list of parameter unit values.}

\item{visit_format}{Type of visit either "WEEK" or "CYCLE".}

\item{n_assessments}{Number of weeks or cycles.}

\item{n_days}{Number of days within cycle.}

\item{seed}{Seed for random number generation.}
}
\value{
data.frame
}
\description{
Function for generating random dataset from Vital Signs Analysis Dataset for a given
Subject-Level Analysis Dataset.
}
\details{
One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID USUBJID PARAMCD AVISITN.
}
\examples{
ADSL <- radsl(seed = 1)
ADVS <- radvs(ADSL, visit_format = "WEEK", n_assessments = 7, seed = 2)
ADVS <- radvs(ADSL, visit_format = "CYCLE", n_assessments = 3, seed = 2)
head(ADVS)

}
\author{
npaszty
}
