% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rel_var}
\alias{rel_var}
\title{Related variables: assign}
\usage{
rel_var(df = NULL, var_values = NULL, related_var = NULL)
}
\arguments{
\item{df}{data frame containing the related variables.}

\item{var_values}{values related to values of `related_var`.}

\item{related_var}{name of variable within `df` with values to which `var_values` must relate.}
}
\description{
Assign values to a related variable within a domain
}
\examples{
# Example with data.frame.
params <- c("Level A", "Level B", "Level C")
ADLB_df <- data.frame(
  ID = 1:9,
  PARAM = factor(
    rep(c("Level A", "Level B", "Level C"), 3),
    levels = params
  )
)
random.cdisc.data:::rel_var(
  df = ADLB_df,
  var_values = c("A", "B", "C"),
  related_var = "PARAM"
)

# Example with tibble.
ADLB_tbl <- tibble::tibble(
  ID = 1:9,
  PARAM = factor(
    rep(c("Level A", "Level B", "Level C"), 3),
    levels = params
  )
)
random.cdisc.data:::rel_var(
  df = ADLB_tbl,
  var_values = c("A", "B", "C"),
  related_var = "PARAM"
)
}
