% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rel_var}
\alias{rel_var}
\title{Related variables: assign}
\usage{
rel_var(df, var_name, related_var, var_values = NULL)
}
\arguments{
\item{df}{data frame containing the related variables.}

\item{var_name}{name of variable related to `rel_var` to add to `df`}

\item{related_var}{name of variable within `df` with values to which values of `var_name` must relate.}

\item{var_values}{values related to values of `related_var`.}
}
\value{
`df` with added factor variable `var_name` containing `var_values` corresponding to `related_var`.
}
\description{
Assign values to a related variable within a domain
}
\examples{
# Example with data.frame.
params <- c("Level A", "Level B", "Level C")
ADLB_df <- data.frame(
  ID = 1:9,
  PARAM = factor(
    rep(c("Level A", "Level B", "Level C"), 3),
    levels = params
  )
)
random.cdisc.data:::rel_var(
  df = ADLB_df,
  var_name = "PARAMCD",
  var_values = c("A", "B", "C"),
  related_var = "PARAM"
)

# Example with tibble.
ADLB_tbl <- tibble::tibble(
  ID = 1:9,
  PARAM = factor(
    rep(c("Level A", "Level B", "Level C"), 3),
    levels = params
  )
)
random.cdisc.data:::rel_var(
  df = ADLB_tbl,
  var_name = "PARAMCD",
  var_values = c("A", "B", "C"),
  related_var = "PARAM"
)
}
