% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rel_var}
\alias{rel_var}
\title{Related variables: assign}
\usage{
rel_var(df = NULL, var_name = NULL, var_values = NULL, related_var = NULL)
}
\arguments{
\item{df}{data frame containing the related variables.}

\item{var_name}{related to variable name.}

\item{var_values}{values of related to variable.}

\item{related_var}{variable name with existing values to which var_name values must relate.}
}
\description{
Assign values to a related variable within a domain
}
\examples{
ADLB <- radlb(radsl())
random.cdisc.data:::rel_var(
  df = ADLB,
  var_name = "PARAMCD",
  var_values = c("ALT", "CRP", "IGA"),
  related_var = "PARAM"
)
}
