% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_na}
\alias{replace_na}
\title{Replace Values in a Vector by NA}
\usage{
replace_na(v, percentage = 0.05, seed = NULL)
}
\arguments{
\item{v}{(\code{any})\cr Vector of any type.}

\item{percentage}{(\code{proportion})\cr Value between 0 and 1 defining
how much of the vector shall be replaced by \code{NA}. This number
is randomized by +/- 5\% to have full randomization.}

\item{seed}{(\code{numeric})\cr Seed to use for reproducible random number generation.}
}
\value{
The input vector \code{v} where a certain number of values are replaced by \code{NA}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Randomized replacement of values by \code{NA}.
}
