% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_na}
\alias{replace_na}
\title{Replace Values in a Vector by NA}
\usage{
replace_na(v, percentage = 0.05, seed = NULL)
}
\arguments{
\item{v}{(`any`)\cr Vector of any type.}

\item{percentage}{(`proportion`)\cr Value between 0 and 1 defining
how much of the vector shall be replaced by `NA`. This number
is randomized by +/- 5\% to have full randomization.}

\item{seed}{(`integer`)\cr Seed for reproducible random number generation.}
}
\value{
The input vector `v` where a certain number of values are replaced by `NA`.
}
\description{
`r lifecycle::badge("stable")`

Randomized replacement of values by `NA`.
}
