% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retain}
\alias{retain}
\title{Retain values within primary keys}
\usage{
retain(df, value_var, event, outside = NA)
}
\arguments{
\item{df}{data frame in which to apply the retain.}

\item{value_var}{variable containing the value to be retained.}

\item{event}{flag to trigger the retain.}

\item{outside}{value to.}
}
\description{
Retain values within primary keys
}
\examples{
ADLB <- radlb(radsl(N = 10, na_percentage = 0), na_vars = list())
ADLB$BASE2 <- random.cdisc.data:::retain(df = ADLB, value_var = ADLB$AVAL,
  event = ADLB$ABLFL2 == "Y")
}
