% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retain}
\alias{retain}
\title{Retain values within primary keys.}
\usage{
retain(df, value_var, event, outside = NA)
}
\arguments{
\item{df}{data frame in which to apply the retain.}

\item{value_var}{variable containing the value to be retained.}

\item{event}{flag to trigger the retain.}

\item{outside}{value to.}
}
\description{
Retain values within primary keys.
}
\examples{
ADLB$BASE2 <- retain(ADLB, ADLB$AVAL, ADLB$ABLFL2 == "Y")

}
