% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_fct}
\alias{sample_fct}
\title{Create a Factor with Random Elements of x}
\usage{
sample_fct(x, N, ...)
}
\arguments{
\item{x}{(`character vector` or `factor`)\cr If character vector then it is also used
as levels of the returned factor. If factor then the levels are used as the new levels.}

\item{N}{Number of items to choose.}

\item{...}{Additional arguments to be passed to `sample`.}
}
\value{
A factor of length `N`.
}
\description{
Sample elements from `x` with replacement to build a factor.
}
\examples{
random.cdisc.data:::sample_fct(letters[1:3], 10)
random.cdisc.data:::sample_fct(iris$Species, 10)
}
\keyword{internal}
