% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_fct}
\alias{sample_fct}
\title{Create a factor with random elements of x}
\usage{
sample_fct(x, N, ...)
}
\arguments{
\item{x}{character vector or factor, if character vector then it is also used
as levels of the returned factor, otherwise if it is a factor then the
levels get used as the new levels}

\item{N}{number of}

\item{...}{arguments passed on to \code{\link{sample}}}
}
\value{
a factor of length N
}
\description{
Sample elements from x with replacing and build a factor
}
\examples{
random.cdisc.data:::sample_fct(letters[1:3], 10)
random.cdisc.data:::sample_fct(iris$Species, 10)
}
\keyword{internal}
