% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_relabel}
\alias{var_relabel}
\title{Apply labels to ADSL primary key variables}
\usage{
var_relabel(x, ...)
}
\arguments{
\item{x}{data frame containing variables to which labels are applied.}

\item{...}{ellipsis.}
}
\description{
Apply labels to ADSL primary key variables
}
\examples{
\dontrun{
 var_relabel(STUDYID = "Study Identifier", USUBJID = "Unique Subject Identifier")
}
}
