% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_relabel}
\alias{var_relabel}
\title{Primary Keys: Labels}
\usage{
var_relabel(x, ...)
}
\arguments{
\item{x}{(`data.frame`)\cr Data frame containing variables to which labels are applied.}

\item{...}{(`named character`)\cr Name-Value pairs, where name corresponds to a variable
name in `x` and the value to the new variable label.}
}
\description{
Relabel a subset of variables in a data set.
}
\examples{
ADSL <- radsl()
random.cdisc.data:::var_relabel(ADSL,
  STUDYID = "Study Identifier",
  USUBJID = "Unique Subject Identifier"
)
}
\keyword{internal}
