% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_relabel}
\alias{var_relabel}
\title{Apply labels to ADSL primary key variables, available in tern.}
\usage{
var_relabel(x, ...)
}
\arguments{
\item{x}{data frame containing variables to which labels are applied.}
}
\description{
Apply labels to ADSL primary key variables, available in tern.
}
\examples{
 var_relabel(
    STUDYID = "Study Identifier",
    USUBJID = "Unique Subject Identifier"
    )

}
