% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_relabel}
\alias{var_relabel}
\title{Primary keys: labels}
\usage{
var_relabel(x, ...)
}
\arguments{
\item{x}{data frame containing variables to which labels are applied.}

\item{...}{ellipsis.}
}
\description{
Apply labels to ADSL primary key variables
}
\examples{
ADSL <- radsl()
random.cdisc.data:::var_relabel(ADSL, STUDYID = "Study Identifier",
USUBJID = "Unique Subject Identifier")
}
