% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{visit_schedule}
\alias{visit_schedule}
\title{Create visit schedule}
\usage{
visit_schedule(visit_format = "WEEK", n_assessments = 10, n_days = 5,
  n_timepoints = NULL, times = NULL, n_timepoints_prior_trt = NULL)
}
\arguments{
\item{visit_format}{as character string. Valid values: WEEK, CYCLE.}

\item{n_assessments}{number of assessments. Valid values: integer.}

\item{n_days}{number of days for each cycle: Valid values: integer.}

\item{n_timepoints}{number of timepoints: Valid values: integer. To Be Implemented.}

\item{times}{nominal time point description as character vector. To Be Implemented.}

\item{n_timepoints_prior_trt}{number of timepoints prior to treatment at visit. Valid values: integer. To Be Implemented.}
}
\value{
a factor of length n_assessments
}
\description{
Create a visit schedule as factor.
}
\details{
X number of visits or X number of cycles and Y number of days.
}
\examples{
\dontrun{
AVISIT <- visit_schedule(visit_format = "WEeK", n_assessments = 10)
AVISIT <- visit_schedule(visit_format = "CyCLE", n_assessments = 5, n_days = 2)
}
}
