% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{visit_schedule}
\alias{visit_schedule}
\title{Create visit schedule}
\usage{
visit_schedule(visit_format = "WEEK", n_assessments = 10L,
  n_days = 5L)
}
\arguments{
\item{visit_format}{as character string. Valid values: WEEK, CYCLE.}

\item{n_assessments}{number of assessments. Valid values: integer.}

\item{n_days}{number of days for each cycle: Valid values: integer.}
}
\value{
a factor of length n_assessments
}
\description{
Create a visit schedule as factor.
}
\details{
X number of visits or X number of cycles and Y number of days.
}
\examples{
random.cdisc.data:::visit_schedule(visit_format = "WEeK", n_assessments = 10L)
random.cdisc.data:::visit_schedule(visit_format = "CyCLE", n_assessments = 5L, n_days = 2L)
}
